Renamed platformio project to PGDrive
