--- 220129-142534-teensy31/src/Samples.h
@@ -1,20 +1,0 @@
-#ifndef Samples_H   // Put these two lines at the top of your file.
-#define Samples_H    // (Use a suitable name, usually based on the file name.)
-
-float realNoise4T[4096] {
-  600.9079855, 620.4374885, 793.9214739, 793.8842376, 629.11232, 676.7649358, 601.7281939, 691.57632, 671.1211442, 671.8025697, 691.8034618, 805.8192291, 805.8192291, 806.9177018, 810.4030255, 804.0008533, 885.4782448, 868.2800097, 861.0077479, 868.5456291, 893.3946958, 921.9562279, 928.6227782, 907.2205576, 879.0760727, 865.2117333, 864.7574497, 870.1740994, 878.0917915, 886.0454788, 889.9850861, 886.4252897, 878.6590255, 874.1137067, 877.7504582, 884.1898667, 885.2126255, 883.5456776, 894.0388848, 923.2061285, 960.3283006, 983.7388024, 980.4433842, 958.2455467, 940.7444558, 941.3129309, 952.7147055, 960.1011588, 956.9956461, 940.4031224, 916.3496727, 903.887903, 919.1908073, 950.6307103, 973.3213091, 980.8976679, 981.1620461, 975.70816, 966.7677091, 967.2604703, 981.7305212, 991.8066812, 980.1392873, 959.4197333, 954.6087952, 969.950177, 986.6556509, 986.995743, 973.2071176, 962.2223903, 968.1702788, 984.5716558, 991.4653479, 979.1550061, 964.4193358, 964.4950497, 972.9799758, 972.9427394, 970.2915103, 982.6403297, 1005.405401, 1013.057474, 995.064863, 966.0478061, 944.7225406, 942.5628315, 965.0250473, 1000.974274, 1027.073241, 1026.126196, 1004.496834, 981.0863321, 974.1541624, 987.0342206, 999.1558982, 985.2915588, 954.6460315, 946.0096776, 974.3068315, 1013.738899, 1031.315704, 1018.966885, 991.0495418, 971.0486497, 972.6014061, 979.7234812, 967.3746618, 940.2901721, 927.7899248, 938.5475103, 954.0030836, 960.2910642, 955.2145067, 940.1387442, 926.9943079, 935.1379006, 961.5782012, 981.23776, 985.1773673, 987.9812655, 991.1252558, 978.5865309, 949.6464291, 929.1155394, 934.0021915, 956.8814545, 976.8438691, 975.1769212, 950.2149042, 922.6376533, 916.9938618, 930.5168679, 941.6542642, 942.4114036, 944.4941576, 949.6079515, 945.7068218, 932.0696242, 924.0774594, 926.8801164, 926.3501188, 914.9483442, 903.2052364, 898.2813479, 897.9400145, 899.8725818, 906.6148461, 915.2896776, 917.4109091, 910.0616921, 890.1749915, 859.1136582, 831.1206012, 828.7337503, 859.1893721, 905.3649455, 939.3803636, 940.0245527, 910.1758836, 879.3789285, 874.3408485, 886.8038594, 892.1063176, 890.2879418, 892.9019345, 890.1365139, 865.666017, 830.2865067, 812.5967515, 823.8855758, 855.2870012, 886.7653818, 895.2887855, 874.1137067, 844.0751321, 826.6125188, 816.9534061, 809.0357139, 814.9836024, 842.4466618, 866.7644897, 863.1289794, 838.0142933, 818.2790206, 819.2260655, 832.2947879, 835.0217309, 812.3696097, 774.8688679, 753.0496, 765.0570861, 796.3840388, 818.7717818, 820.2103467, 811.0812315, 805.9302012, 802.7477333, 793.5813818, 780.5126594, 775.2846739, 781.4212267, 793.2785261, 804.6803006, 810.3240921, 806.1188655, 796.8010861, 791.2330085, 785.0964558, 765.2469915, 736.3826036, 721.6469333, 729.4119564, 743.6548655, 747.4058085, 742.6321067, 736.4583176, 732.9357576, 736.3441261, 750.6255127, 767.6338424, 771.4592582, 757.7463467, 741.0793503, 735.3213673, 741.6850618, 756.7620655, 773.4662982, 781.0811345, 776.6102885, 776.9900994, 796.9525139, 820.2103467, 817.4449261, 790.9673891, 766.6110836, 763.7314715, 780.5126594, 803.8462061, 814.6422691, 799.2624097, 771.8005915, 755.3979733, 754.56512, 758.3532994, 766.3827006, 785.3993115, 811.3083733, 834.7188752, 849.1504485, 846.2336, 825.0212848, 803.6190642, 802.7862109, 819.7175855, 840.9311418, 859.9465115, 870.5911467, 865.9316364, 853.9241503, 854.4913842, 867.4843927, 875.3263709, 872.9010424, 868.3184873, 865.4773527, 858.6966109, 847.9762618, 837.1044848, 832.8620218, 843.1280873, 864.4161164, 879.3032145, 875.250657, 866.0830642, 870.0983855, 882.3342545, 883.4699636, 869.9469576, 856.4996655, 855.81824, 862.7504097, 864.0003103, 849.5302594, 826.2339491, 813.4296048, 822.4457697, 840.8169503, 842.1810424, 822.4457697, 805.5516315, 810.2099006, 826.4610909, 831.914977, 823.6571927, 816.4978812, 822.7101479, 839.8326691, 854.4156703, 858.9994667, 858.8480388, 863.9618327, 869.3784824, 867.3714424, 862.4847903, 869.5299103, 889.8721358, 916.8424339, 938.206177, 939.0775079, 913.5085382, 885.4025309, 888.3193794, 919.9864242, 947.3737697, 951.9948024, 942.2599758, 928.9641115, 916.3124364, 910.5159758, 910.933023, 912.7886352, 916.046817, 926.047263, 932.4866715, 922.8647952, 904.5693285, 892.7132703, 889.9850861, 887.3723345, 883.9999612, 884.5312, 892.7889842, 911.3115927, 934.0779055, 947.6009115, 943.4713988, 929.7212509, 915.5168194, 902.1067636, 890.3264194, 885.742623, 889.8336582, 895.6301188, 891.57632, 876.1977018, 859.7950836, 856.0069042, 861.34784, 857.6738521, 838.1272436, 821.4230109, 828.0895612, 854.0755782, 872.7123782, 867.2957285, 841.8397091, 820.5889164, 819.3017794, 837.7871515, 856.6895709, 860.4007952, 843.5823709, 818.2790206, 799.7179345, 794.6041406, 798.4295564, 804.2247758, 810.7411394, 816.1950255, 817.3692121, 811.1197091, 797.3695612, 781.9139879, 773.1634424, 773.6946812, 775.20896, 771.7248776, 768.9209794, 774.4518206, 783.4295079, 790.1345358, 795.2483297, 799.7936485, 797.5209891, 784.2251248, 763.4670933, 748.5415176, 751.6482715, 772.1791612, 790.9289115, 795.6641358, 795.6641358, 802.7862109, 810.8540897, 814.0365576, 817.4076897, 819.7945406, 810.0969503, 794.5284267, 791.4601503, 801.0807855, 806.3844848, 805.1345842, 807.9757188, 812.5967515, 812.1424679, 808.6571442, 807.44448, 805.4759176, 804.5288727, 812.8996073, 826.385377, 832.14336, 827.6725139, 824.3398594, 827.2182303, 832.2563103, 834.2645915, 831.839263, 824.5670012, 815.2479806, 810.8168533, 815.0965527, 827.4081358, 846.1578861, 869.1513406, 892.0690812, 907.5234133, 912.7513988, 917.7510012, 930.1755345, 943.1313067, 940.1387442, 919.2292848, 896.1216388, 884.4554861, 884.8340558, 887.9395685, 883.9627248, 871.6139055, 860.627937, 867.3329648, 892.7505067, 921.8420364, 935.6306618, 928.5470642, 907.5618909, 882.3714909, 858.545183, 841.3854255, 841.0068558, 861.3863176, 887.4852848, 896.1601164, 880.704543, 852.5985358, 828.8467006, 818.0134012, 819.5661576, 821.3845333, 814.9078885, 807.5586715, 809.679903, 812.3311321, 799.1866958, 773.5047758, 756.6106376, 765.0955636, 795.5884218, 828.8851782, 841.991137, 829.6050812, 807.3315297, 791.5731006, 782.3310352, 779.2627588, 788.1647321, 808.3915248, 824.794143, 821.1573915, 795.7398497, 760.09472, 740.1323055, 753.5038836, 790.2102497, 815.0208388, 808.8842861, 785.9293091, 772.2163976, 779.0728533, 796.4597527, 807.1801018, 802.634783, 790.0203442, 781.6483685, 776.2701964, 765.1712776, 747.4430448, 732.9357576, 731.5716655, 743.5419152, 760.4745309, 770.3223079, 767.5581285, 761.3830982, 766.7240339, 786.3835927, 809.9827588, 824.3770958, 831.2720291, 837.7486739, 844.8707491, 844.3779879, 835.5517285, 829.1507976, 836.9158206, 860.2493673, 891.5390836, 915.0997721, 917.2210036, 896.2730667, 871.6511418, 861.4992679, 867.181537, 880.9701624, 897.6371588, 914.6070109, 928.6227782, 939.5702691, 952.4490861, 968.0560873, 979.2679564, 979.4206255, 972.3742642, 970.4429382, 979.9878594, 990.5567806, 986.3515539, 969.2687515, 957.2985018, 968.2832291, 997.7918061, 1028.133236, 1040.293392, 1029.687234, 1006.80673, 985.78432, 975.9725382, 982.9804218, 1005.329687, 1027.754667, 1031.23999, 1013.625949, 988.4727855, 966.843423, 953.8131782, 955.7829818, 971.2385552, 983.2832776, 977.4123442, 960.9340121, 941.5028364, 914.7212024, 879.1517867, 856.5753794, 867.9399176, 902.372383, 927.1842133, 920.3277576, 887.7509042, 851.0073018, 829.6050812, 831.3092655, 849.1889261, 867.2200145, 867.4471564, 842.71104, 799.3393648, 759.9060558, 747.6701867, 768.6566012, 800.2094545, 809.1499055, 785.5507394, 747.6329503, 720.81408, 712.8206739, 719.9799855, 736.3826036, 750.814177, 750.5113212, 735.8526061, 725.056543, 729.8290036, 738.9208824, 737.6697406, 728.9204364, 728.654817, 736.8368873, 736.0412703, 715.8132364, 690.0171248, 684.9790448, 709.4495418, 747.5944727, 771.8005915, 768.0496485, 743.0876315, 714.8661915, 701.3059491, 710.2836364, 736.9883152, 762.3673794, 763.8828994, 734.8298473, 694.4867297, 673.9943176, 692.2525479, 740.9279224, 790.6645333, 810.7026618, 794.565663, 766.8754618, 761.9888097, 789.4146327, 825.3626182, 844.9079855, 847.1049309, 844.8322715, 840.3254303, 828.1280388, 817.9761648, 827.0668024, 853.4686255, 871.9924752, 866.1587782, 847.5604558, 836.9915345, 835.5902061, 832.0291685, 823.0142448, 817.7105455, 820.7403442, 826.309663, 826.801183, 824.3013818, 827.4453721, 838.8471467, 850.0217794, 846.953503, 830.4006982, 817.4076897, 825.8926158, 847.5604558, 859.1136582, 851.1959661, 842.2952339, 851.3858715, 871.8025697, 880.0231176, 866.9928727, 852.1430109, 858.7723248, 888.6594715, 914.6827248, 912.6372073, 884.6826279, 856.7268073, 847.0676945, 852.5228218, 859.6051782, 862.5605042, 860.85632, 852.5985358, 838.6957188, 828.1652752, 829.3394618, 840.6282861, 854.4541479, 865.666017, 869.41696, 861.9920291, 845.6278885, 834.6803976, 837.8628655, 848.469023, 850.2116848, 840.1355248, 829.907937, 831.914977, 844.0366545, 853.9998642, 853.3929115, 848.1661673, 847.8633115, 852.6742497, 849.8703515, 838.0142933, 832.2947879, 848.2034036, 876.6519855, 894.9859297, 896.5014497, 896.3115442, 907.7517964, 926.6529745, 941.3501673, 947.48672, 944.759777, 932.4109576, 913.0927321, 898.12992, 897.2585891, 907.2205576, 913.0170182, 908.5461721, 903.2052364, 913.5842521, 940.4031224, 965.6692364, 971.465697, 958.4342109, 947.1466279, 952.6762279, 971.2757915, 985.0259394, 986.5414594, 985.1773673, 995.7847661, 1019.989644, 1045.710041, 1062.869799, 1070.180538, 1073.703098, 1074.347287, 1073.286051, 1072.718817, 1074.498715, 1074.005954, 1064.801125, 1050.331074, 1039.497775, 1038.28511, 1041.391864, 1042.376145, 1038.512252, 1032.527127, 1028.020286, 1027.944572, 1031.921416, 1037.527971, 1041.467578, 1043.21024, 1044.043093, 1045.747278, 1046.657086, 1041.201959, 1027.110478, 1012.488999, 1011.920524, 1027.71743, 1044.270235, 1043.626046, 1027.300383, 1013.018996, 1013.474521, 1025.557721, 1036.27807, 1033.13408, 1012.148907, 984.0416582, 967.2989479, 971.6543612, 988.5484994, 1004.04255, 1013.36033, 1014.800136, 1004.572548, 982.8674715, 961.3510594, 951.8818521, 952.5248, 953.6232727, 955.4416485, 968.9646545, 1000.974274, 1039.80063, 1065.862361, 1070.824727, 1068.323685, 1078.551273, 1103.325867, 1123.363995, 1123.476945, 1112.909265, 1106.696999, 1104.120242, 1094.158274, 1079.688223, 1074.764335, 1082.983641, 1090.56, 1084.840495, 1076.127185, 1087.453246, 1123.476945, 1157.645033, 1159.463408, 1130.826162, 1105.105765, 1107.984136, 1133.515869, 1155.902371, 1164.387297, 1168.858143, 1178.478778, 1184.009619, 1174.350507, 1152.834095, 1132.000349, 1120.977144, 1119.727244, 1126.925033, 1137.871282, 1145.144785, 1140.409561, 1125.068179, 1112.682124, 1118.060296, 1139.728136, 1163.744349, 1175.827549, 1177.07745, 1181.661246, 1201.358041, 1226.169872, 1227.988247, 1197.608339, 1159.766264, 1143.592029, 1148.630109, 1151.698385, 1137.53119, 1113.742119, 1101.241872, 1112.567932, 1141.508034, 1169.009571, 1179.994298, 1175.637644, 1170.940897, 1180.638487, 1197.722531, 1202.229372, 1187.192087, 1170.108044, 1167.607001, 1180.411345, 1196.888436, 1206.093265, 1204.086225, 1198.025387, 1197.873959, 1203.327845, 1206.510313, 1201.547947, 1191.661692, 1186.660848, 1199.312524, 1225.715588, 1240.375544, 1226.548441, 1201.206613, 1192.987307, 1206.017552, 1221.548839, 1227.19263, 1223.783021, 1215.86657, 1202.570705, 1182.872669, 1157.569319, 1131.470352, 1114.689164, 1115.711922, 1130.901876, 1139.500994, 1119.614293, 1074.537193, 1035.443976, 1032.679796, 1061.732848, 1084.764781, 1067.453595, 1011.617668, 949.6464291, 910.3645479, 895.9702109, 891.2734642, 887.5237624, 888.2051879, 891.3491782, 879.6445479, 840.5525721, 788.8833939, 760.7389091, 774.9445818, 812.6352291, 837.5600097, 833.6961164, 816.233503, 808.6956218, 813.6952242, 817.0663564, 810.1341867, 803.1647806, 805.0588703, 812.5582739, 811.8396121, 798.5809842, 777.7087612, 760.09472, 757.3305406, 768.4282182, 779.7170424, 781.3082764, 780.6256097, 782.3682715, 772.2548752, 738.4653576, 699.7519515, 690.6228364, 722.4040727, 766.0798448, 779.451423, 753.8067394, 727.0263467, 733.5414691, 762.4815709, 785.1336921, 804.8689648, 836.4230594, 858.3180412, 836.6502012, 779.7170424, 739.45088, 748.1257115, 780.9297067, 788.0505406, 762.405857, 752.6325527, 785.7778812, 817.4449261, 791.6115782, 720.9655079, 667.2508121, 663.4253964, 683.236383, 673.5400339, 612.0987927, 539.5958691, 514.5196606, 540.3914861, 568.2715927, 567.8917818, 555.1271564, 554.0286836, 564.6720776, 560.0138085, 521.6032582, 467.6998982, 439.21408, 446.563297, 470.0110352, 502.0566497, 545.4295661, 572.2484364, 538.4216824, 443.2294012, 350.4996848, 329.0589867, 378.4927418, 426.0696436, 404.6301867, 328.7946085, 269.7017406, 272.8072533, 320.9911079, 371.5220945, 393.7956461, 379.061217, 336.1053479, 280.1564703, 231.4810958, 210.4946812, 216.4425697, 222.0106473, 209.2832582, 203.1467055, 226.9730133, 263.90528, 284.4361697, 280.4208485, 257.0115879, 227.2758691, 219.3594182, 249.2080873, 290.3840582, 296.8991806, 254.7389285, 194.2447321, 158.5238885, 164.0919661, 197.6915782, 239.5862109, 275.6868655, 290.3083442, 276.1026715, 254.9288339, 267.8833648, 345.2716994, 474.6320679, 582.9687855, 581.3403152, 475.8434909, 382.9623467, 389.2130909, 448.6460509, 479.0644364, 470.4268412, 463.0031515, 479.0644364, 512.1328097, 537.4746376, 559.4453333, 624.9776097, 742.5948703, 817.0663564, 763.0115685, 629.5614061, 544.9752824, 572.7784339, 681.8735321, 803.9603976, 873.9622788, 844.4164655, 738.3139297, 670.8118497, 727.5935806, 846.953503, 917.6380509, 939.1532218, 963.396577, 980.4049067, 971.5414109, 958.5099248, 942.8656873, 911.0844509, 890.8948945, 907.3347491, 938.3203685, 970.67008, 1019.004121, 1066.770928, 1072.528912, 1034.459695, 1003.398361, 1026.997527, 1104.15872, 1185.108092, 1204.047748, 1129.273406, 1017.602793, 987.7541236, 1087.831816, 1227.07968, 1299.922696, 1305.415059, 1292.232145, 1280.868848, 1281.85313, 1312.991418, 1366.136398, 1411.32769, 1417.767098, 1359.658512, 1224.351496, 1066.582264, 997.4132364, 1083.665067, 1257.42111, 1386.667287, 1398.826201, 1322.612053, 1249.844752, 1261.663573, 1344.241416, 1405.758371, 1405.985513, 1394.129455, 1404.546948, 1401.781527, 1339.317527, 1228.102439, 1138.478235, 1123.401232, 1155.902371, 1180.145726, 1175.410502, 1147.151825, 1131.924635, 1183.25248, 1261.057862, 1258.898153, 1199.389479, 1215.829333, 1317.11969, 1377.425222, 1324.051859, 1187.607893, 1054.839156, 1031.088562, 1155.562279, 1344.355607, 1449.396907, 1375.644082, 1174.653362, 1020.216785, 1010.103389, 1073.437479, 1139.9168, 1223.518642, 1283.254458, 1219.200465, 1040.709198, 897.7513503, 932.9794327, 1149.084393, 1394.129455, 1492.655632, 1380.796354, 1123.515423, 898.6983952, 920.3277576, 1200.866521, 1463.791244, 1452.048136, 1211.245537, 983.7388024, 964.7606691, 1163.895777, 1401.630099, 1467.048184, 1323.104815, 1109.234036, 981.7305212, 987.6026958, 1041.997576, 1042.263195, 991.2009697, 936.3890424, 856.8025212, 745.3602909, 693.5781624, 777.2544776, 937.9790352, 1045.180044, 1012.376048, 863.3176436, 724.9808291, 734.2613721, 915.0997721, 1139.349566, 1226.132635, 1110.370987, 916.1982448, 818.1275927, 881.7273018, 1040.103486, 1155.789421, 1112.870788, 931.1982933, 761.1174788, 718.8442764, 778.8084752, 843.6580848, 849.984543, 793.3157624, 714.2220024, 665.0166303, 664.9409164, 688.5400824, 713.464863, 750.2084655, 782.1796073, 753.4281697, 668.91776, 630.6598788, 687.4031321, 749.9428461, 713.6547685, 583.9915442, 454.2526061, 423.8354618, 518.0422206, 654.523423, 710.2836364, 646.1142109, 534.9363588, 482.1327127, 523.0803006, 605.3180509, 659.5999806, 674.7142206, 670.0919467, 645.24288, 574.7482376, 460.6920145, 361.4086982, 337.5439127, 367.5080145, 374.5903709, 331.5215515, 274.8155345, 245.6855273, 282.770463, 394.175457, 501.8295079, 512.1712873, 447.0932945, 396.0683055, 386.82624, 392.1299394, 402.3190497, 408.0770327, 387.8489988, 365.1211636, 393.87136, 452.2070885, 447.9261479, 367.5452509, 318.9455903, 372.3934255, 467.2828509, 510.7699588, 491.5646836, 458.6849745, 455.8053624, 485.8824145, 511.9056679, 496.7157139, 433.456097, 337.0139152, 236.4049842, 186.5926594, 220.7607467, 298.4904145, 348.03712, 357.8861382, 367.8865842, 390.3488, 387.7348073, 335.5368727, 265.9507976, 233.5638497, 271.5958303, 363.4169794, 434.8201891, 403.3418085, 283.3004606, 188.9795103, 194.5860655, 264.3223273, 334.8554473, 379.552737, 390.6516558, 378.4170279, 376.3330327, 392.6214594, 401.4861964, 401.3720048, 417.774623, 447.8131976, 460.3891588, 447.4718642, 424.1755539, 410.5768339, 434.4031418, 511.4886206, 618.4240097, 706.911263, 741.3064921, 707.5182158, 629.333023, 576.3779491, 602.9312, 671.3418473, 690.2442667, 639.8647079, 598.4615952, 634.7509139, 719.3357964, 779.4141867, 797.4067976, 819.4904436, 880.2502594, 956.5785988, 990.7839224, 962.1851539, 925.6302158, 954.9116509, 1040.974817, 1092.869896, 1035.861023, 890.780703, 751.9126497, 717.7073261, 814.2252218, 969.0403685, 1080.635268, 1113.022216, 1102.6432, 1087.264582, 1077.681183, 1080.066793, 1104.840145, 1147.304495, 1186.509421, 1198.062623, 1159.236267, 1077.908325, 1015.217183, 1034.194075, 1121.811239, 1209.995636, 1253.974264, 1257.004063, 1240.526972, 1231.132238, 1249.466182, 1289.391011, 1315.339792, 1298.899937, 1249.011898, 1201.206613, 1179.652965, 1179.274395, 1184.463903, 1192.760165, 1208.517353, 1229.616718, 1243.329629, 1249.124848, 1263.823282, 1293.293382, 1308.256194, 1281.740179, 1234.314705, 1207.911641, 1216.283617, 1237.988693, 1245.450861, 1229.806623, 1207.684499, 1201.169377, 1212.874007, 1230.109479, 1250.86751, 1284.542836, 1328.672892, 1359.924131, 1354.01472, 1316.892548, 1287.535399, 1299.315743, 1342.651423, 1372.045808, 1357.082996, 1319.392349, 1313.142846, 1367.916296, 1453.600892, 1509.474056, 1496.898095, 1433.789905, 1378.030933, 1374.886943, 1421.13823, 1472.806167, 1484.322133, 1437.00961, 1354.431767, 1292.573479, 1299.922696, 1381.477779, 1488.716024, 1554.778298, 1546.217658, 1587.768979, 1539.357983, 1546.82015, 1614.512136, 1700.938996, 2300.441474, 2060.085217, 1883.264116, 1697.920853, 1566.593901, 1492.54144, 1517.997459, 1499.625038, 1459.093256, 1445.911583, 1461.707248, 1455.533459, 1395.796402, 1323.483384, 1314.771316, 1389.28128, 1482.08671, 1511.253954, 1460.647253, 1384.546056, 1342.158662, 1349.620829, 1382.576252, 1412.841968, 1429.547442, 1437.653799, 1438.070846, 1412.917682, 1345.188461, 1252.64865, 1195.032824, 1220.790458, 1309.391903, 1378.712359, 1357.992805, 1255.602735, 1146.546114, 1102.908819, 1133.932916, 1187.759321, 1207.798691, 1177.228878, 1122.871234, 1083.362211, 1074.30881, 1078.211181, 1065.483792, 1034.611122, 1012.640427, 1021.693828, 1047.187084, 1055.559059, 1035.179598, 1002.52703, 981.0106182, 979.7991952, 994.7620073, 1015.898608, 1030.558565, 1026.883336, 996.9204752, 947.7523394, 906.3492267, 906.4634182, 952.03328, 1005.783971, 1015.898608, 961.5024873, 874.5679903, 822.9000533, 853.9998642, 943.0555927, 1005.102545, 979.1165285, 894.60736, 835.0974448, 846.6891248, 897.6756364, 926.2371685, 901.5382885, 843.0523733, 791.9516703, 774.8303903, 785.0207418, 797.2938473, 793.2400485, 773.5804897, 752.4811248, 746.4587636, 758.5047273, 772.6719224, 768.9209794, 750.0942739, 736.7984097, 736.8753648, 736.1926982, 724.0337842, 713.5033406, 727.2150109, 767.5581285, 805.0203927, 801.3464048, 749.6399903, 693.1238788, 686.3046594, 735.1699394, 787.8991127, 792.2172897, 742.4806788, 673.8044121, 626.1517964, 617.7425842, 642.2887952, 674.4101236, 689.5243636, 681.1908655, 662.1370182, 654.107617, 670.43328, 699.3361455, 713.4276267, 700.0932848, 681.6079127, 684.1461915, 703.2372752, 714.5260994, 707.5554521, 694.0324461, 690.8127418, 710.5480145, 751.6097939, 791.7630061, 801.4978327, 773.1634424, 724.2609261, 673.8428897, 634.5225309, 626.8716994, 663.7282521, 722.1769309, 748.8443733, 711.3821091, 632.6669188, 568.1958788, 551.9074521, 571.415583, 587.4768679, 578.0064194, 553.4217309, 533.9520776, 526.4899103, 533.2706521, 561.4908509, 609.2576582, 651.6822885, 664.5238691, 648.5382982, 624.3334206, 607.7421382, 603.9924364, 611.1889842, 625.810463, 642.2130812, 654.3719952, 656.6446545, 652.5536194, 654.523423, 664.5623467, 665.0166303, 646.6826861, 627.1360776, 628.5758836, 639.3334691, 631.1513988, 594.7118933, 546.8308945, 510.3144339, 502.9664582, 530.0497067, 572.5140558, 595.0519855, 585.1657309, 565.0518885, 561.4523733, 574.7482376, 596.681697, 625.09056, 652.6293333, 661.6827345, 642.7430788, 606.8708073, 573.8769067, 560.7709479, 572.9683394, 593.9162764, 601.1128242, 591.2265697, 583.9530667, 587.097057, 590.3552388, 584.5600194, 571.982817, 549.2549818, 508.951583, 466.5629479, 456.4880291, 484.8211782, 510.8456727, 494.7459103, 450.1988073, 424.3269818, 438.6456048, 477.7375806, 516.7538424, 542.8155733, 548.8391758, 535.9591176, 527.2855273, 547.8536533, 591.33952, 620.05248, 615.2043055, 593.1206594, 570.8471079, 547.7779394, 532.3236073, 539.7857745, 567.8160679, 591.4152339, 601.4913939, 608.2348994, 612.7802182, 603.4996752, 579.2947976, 555.0514424, 534.5962667, 507.7773964, 477.0561552, 461.7532509, 471.7524558, 489.6705939, 492.5489648, 469.1397042, 425.7283103, 385.5006255, 370.6135273, 379.5912145, 394.7811685, 407.1299879, 422.0543224, 437.6613236, 444.7064436, 438.3812267, 420.4258521, 396.9396364, 379.477023, 378.6056921, 387.8105212, 390.5759418, 381.8638739, 378.9097891, 401.5234327, 448.6088145, 496.9813333, 519.6706909, 511.3756703, 486.7909818, 461.4876315, 440.7668364, 432.66048, 445.1234909, 467.6614206, 476.0333964, 460.6547782, 436.9414206, 419.2131879, 413.8722521, 424.9339345, 453.4185115, 487.5493624, 509.6330085, 516.4509867, 514.4054691, 511.0728145, 510.0872921, 521.4518303, 552.8160194, 590.7338085, 609.8633697, 598.9543564, 572.2869139, 554.785823, 569.4457794, 618.3482958, 668.5776679, 679.5623952, 650.6222933, 619.5981964, 618.6896291, 651.1522909, 699.6390012, 730.9659539, 717.78304, 667.1378618, 618.04544, 605.4310012, 632.5154909, 677.7440194, 715.4731442, 725.3593988, 704.4871758, 673.7286982, 664.9781527, 689.1085576, 723.3511176, 745.0189576, 755.5878788, 771.0049745, 795.1341382, 817.5963539, 822.3700558, 807.748577, 792.4829091, 800.2094545, 838.6957188, 902.5610473, 973.5869285, 1026.618958, 1049.498221, 1057.226007, 1063.816844, 1065.483792, 1063.286846, 1074.537193, 1105.371384, 1138.099665, 1144.993358, 1121.735525, 1093.476848, 1095.256747, 1135.409959, 1194.842919, 1242.912582, 1252.7616, 1222.15455, 1185.979423, 1183.895428, 1221.586075, 1277.194861, 1323.104815, 1336.51487, 1328.975748, 1345.719699, 1407.91808, 1472.351884, 1487.50336, 1448.298434, 1392.083937, 1361.969648, 1385.683006, 1449.131287, 1506.444257, 1515.876228, 1472.920359, 1413.183302, 1384.697484, 1407.425319, 1456.744882, 1499.549324, 1522.76992, 1528.830759, 1520.155927, 1502.844742, 1486.746221, 1477.768533, 1473.828926, 1467.389518, 1463.184291, 1471.746172, 1488.186027, 1490.571636, 1471.366361, 1438.14656, 1396.3264, 1353.977484, 1333.8624, 1348.143787, 1382.917585, 1408.145222, 1403.561425, 1369.090482, 1331.059743, 1315.339792, 1331.059743, 1368.067724, 1406.895321, 1420.986802, 1402.387239, 1370.984572, 1346.286933, 1335.036587, 1337.347724, 1344.924082, 1340.52895, 1328.408514, 1322.839195, 1314.202841, 1284.657028, 1246.322192, 1221.056078, 1207.911641, 1200.563665, 1204.237653, 1218.101993, 1225.449968, 1214.086672, 1191.547501, 1178.934303, 1183.101052, 1184.615331, 1165.031486, 1132.038827, 1107.68128, 1099.613401, 1105.257193, 1118.211724, 1136.091384, 1156.319418, 1165.410056, 1145.750497, 1097.60512, 1045.21728, 1011.541954, 1006.42816, 1022.603636, 1037.225115, 1036.012451, 1027.261905, 1013.625949, 981.9961406, 941.8814061, 927.8656388, 953.2831806, 1001.201416, 1037.717876, 1021.92097, 941.3501673, 848.9617842, 811.6497067, 834.9075394, 879.3032145, 908.6603636, 904.1522812, 867.9771539, 828.7709867, 808.3158109, 803.4676364, 807.5971491, 819.0746376, 833.7333527, 849.3018764, 857.5596606, 840.8554279, 800.6265018, 765.8911806, 749.7541818, 738.6924994, 731.6846158, 735.8526061, 739.45088, 727.8592, 704.6386036, 675.5855515, 648.8039176, 643.8043152, 655.6603733, 646.5684945, 601.6428218, 558.8011442, 555.4300121, 583.0444994, 607.5907103, 607.6279467, 587.9683879, 567.9674958, 562.0208485, 568.9145406, 574.21824, 563.7250327, 540.9599612, 528.7998061, 543.9140461, 577.8549915, 615.5828752, 641.3032727, 631.0384485, 573.3096727, 498.6098036, 450.7300461, 438.9869382, 449.0630982, 468.0412315, 479.8215758, 466.2985697, 446.563297, 454.5939394, 489.8220218, 517.3607952, 523.5730618, 521.3004024, 515.0124218, 500.4654158, 481.0714764, 465.3515248, 458.3436412, 464.4801939, 476.4119661, 471.6767418, 448.0018618, 434.5930473, 449.6688097, 468.722657, 461.1090618, 432.7746715, 413.7965382, 417.6604315, 432.7746715, 447.4718642, 464.1773382, 482.8513745, 497.9283782, 516.4509867, 547.8921309, 579.4462255, 595.2034133, 606.2266182, 625.3561794, 632.5539685, 607.7036606, 563.2335127, 525.9586715, 513.7240436, 533.8378861, 576.1495661, 610.0532752, 618.5382012, 618.0069624, 633.9925333, 669.7890909, 712.6692461, 746.7616194, 755.7393067, 734.7156558, 711.533537, 725.2079709, 771.3078303, 807.2930521, 817.2562618, 819.6418715, 818.1275927, 807.44448, 813.3923685, 853.582817, 892.9019345, 898.6599176, 896.5386861, 920.5921358, 959.4954473, 975.8968242, 954.2674618, 918.3194764, 918.3194764, 972.4127418, 1036.618162, 1055.18049, 1019.231263, 961.2753455, 933.0551467, 971.4272194, 1053.134972, 1106.16576, 1099.764829, 1069.082065, 1058.096097, 1077.4528, 1112.265076, 1136.357004, 1143.289173, 1159.841978, 1201.055185, 1240.375544, 1251.360272, 1239.617164, 1220.866172, 1205.980315, 1206.813168, 1220.828936, 1238.101644, 1253.064456, 1263.292044, 1263.97471, 1269.428596, 1299.959932, 1343.181421, 1371.515811, 1378.409503, 1374.432659, 1360.94689, 1342.764373, 1334.355161, 1339.583147, 1342.007234, 1331.211171, 1324.393193, 1342.347326, 1373.674279, 1389.39423, 1392.046701, 1403.978473, 1414.281775, 1393.978027, 1352.196344, 1321.020819, 1313.33151, 1327.574419, 1368.674676, 1424.395171, 1458.108975, 1438.335224, 1371.288669, 1299.695554, 1276.587908, 1313.899985, 1361.704029, 1366.326303, 1327.612897, 1283.254458, 1257.042541, 1248.480659, 1250.413227, 1261.019384, 1279.656184, 1303.066686, 1319.316635, 1308.937619, 1262.876238, 1208.365925, 1193.138735, 1226.245585, 1253.89855, 1222.078836, 1149.88001, 1108.476897, 1136.054148, 1195.562822, 1220.980364, 1184.729522, 1117.416107, 1067.073784, 1056.089057, 1074.537193, 1098.930735, 1114.954783, 1120.332955, 1113.514977, 1085.14335, 1035.748073, 995.1778133, 999.7988461, 1052.300878, 1108.779753, 1112.416504, 1055.67201, 992.9051539, 976.8823467, 990.102497, 978.1322473, 933.3580024, 899.9098182, 907.9032242, 940.8958836, 960.2525867, 941.6542642, 897.6371588, 864.0760242, 856.613857, 851.6514909, 829.7565091, 807.44448, 814.1122715, 847.6734061, 872.2580945, 851.1202521, 786.7249261, 726.2307297, 715.8889503, 747.291617, 771.3835442, 762.2916655, 743.9961988, 742.4806788, 742.0648727, 713.2749576, 666.7208145, 638.8791855, 634.5610085, 621.4922861, 588.8397188, 566.6418812, 581.7945988, 618.765343, 638.19776, 614.7872582, 561.7937067, 520.2776436, 522.9673503, 558.6869527, 588.1198158, 582.6659297, 552.3617358, 526.4899103, 521.110497, 526.3000048, 526.6413382, 521.7919224, 521.9818279, 527.7770473, 528.4597139, 517.8907927, 503.9507394, 498.5725673, 499.2539927, 491.3747782, 469.7838933, 445.9563442, 434.7059976, 439.3282715, 458.8364024, 478.7603394, 476.8290133, 446.7519612, 416.9405285, 420.6145164, 455.9195539, 487.2080291, 481.9800436, 446.4118691, 419.7059491, 428.7965867, 450.8429964, 450.8814739, 430.0092509, 421.372897, 438.5326545, 463.608863, 475.2377794, 463.5331491, 432.5847661, 404.2503758, 408.265697, 443.8351127, 470.8066521, 460.7677285, 436.6758012, 442.6236897, 479.2915782, 507.0574933, 496.1100024, 456.9423127, 419.2889018, 403.456, 413.0393988, 438.9112242, 467.0941867, 485.0110836, 492.5117285, 493.3830594, 488.3449794, 479.4045285, 483.3826133, 516.9065115, 569.3315879, 610.5075588, 622.8936145, 614.7500218, 606.8323297, 615.4699248, 639.484897, 658.4630303, 662.8953988, 670.2061382, 702.8959418, 754.1853091, 794.8685188, 809.9455224, 809.7183806, 812.8996073, 821.2331055, 826.7254691, 831.423457, 849.0374982, 877.3334109, 896.4629721, 899.3041067, 904.455137, 926.4643103, 952.2976582, 959.4197333, 945.7068218, 931.6910545, 940.5557915, 977.2609164, 1018.588315, 1026.618958, 994.9506715, 965.0250473, 978.510817, 1021.505164, 1050.596693, 1053.2864, 1047.945464, 1044.232999, 1039.383583, 1035.936737, 1043.21024, 1064.119699, 1094.121038, 1116.12897, 1110.105367, 1079.726701, 1056.733246, 1070.256252, 1115.523258, 1163.440252, 1185.410948, 1171.357944, 1130.333401, 1086.468965, 1071.430439, 1098.893498, 1144.500596, 1167.986812, 1153.668189, 1119.878672, 1092.491326, 1084.082114, 1095.105319, 1114.310594, 1125.219607, 1124.386754, 1121.962667, 1130.333401, 1152.531239, 1182.532577, 1206.775932, 1213.215341, 1203.480514, 1194.881396, 1202.609183, 1219.464844, 1224.42721, 1208.9344, 1186.245042, 1170.296708, 1164.122919, 1167.494051, 1179.085731, 1187.607893, 1180.675724, 1159.5776, 1140.485275, 1135.674337, 1145.674783, 1157.379413, 1156.356655, 1138.251093, 1112.909265, 1096.506647, 1099.12064, 1113.060693, 1115.523258, 1096.733789, 1076.430041, 1076.014235, 1085.370492, 1080.369648, 1063.36256, 1055.937629, 1060.255806, 1057.756005, 1045.520136, 1041.353387, 1054.536301, 1074.460238, 1090.257144, 1099.877779, 1100.067685, 1094.08256, 1093.70399, 1100.333304, 1097.1136, 1074.271573, 1046.505658, 1029.762948, 1024.649154, 1030.10304, 1047.376989, 1069.877682, 1082.870691, 1077.377086, 1058.058861, 1038.853585, 1034.269789, 1045.482899, 1051.695166, 1030.974371, 993.7392485, 976.995297, 998.7760873, 1028.701712, 1025.899055, 993.4736291, 974.8368291, 993.3222012, 1021.996684, 1024.762104, 998.4732315, 963.3593406, 931.161057, 911.6144485, 919.8349964, 953.244703, 977.8666279, 964.2306715, 921.2363248, 877.2961745, 853.1285333, 853.4686255, 873.0152339, 891.5006061, 889.6065164, 867.8642036, 841.8781867, 822.0672, 811.3083733, 814.2636994, 831.2720291, 847.5976921, 836.6129648, 791.118817, 737.6325042, 711.4950594, 718.4657067, 736.1926982, 744.0719127, 736.8753648, 719.9042715, 700.5103321, 688.1230352, 684.3348558, 688.0100848, 697.8578618, 705.3212703, 693.0854012, 658.2358885, 627.5158885, 626.0376048, 648.8039176, 671.1904194, 680.6993455, 672.2516558, 648.0840145, 621.2266667, 614.6358303, 627.7430303, 636.1894788, 626.9474133, 615.318497, 612.3631709, 608.3478497, 591.9079952, 571.8698667, 559.9753309, 556.6414352, 556.3385794, 560.0510448, 573.422623, 590.7338085, 593.1578958, 571.8698667, 545.2781382, 538.0431127, 547.2094642, 556.1871515, 562.7407515, 569.6344436, 561.3766594, 528.9897115, 491.2605867, 475.6163491, 482.8141382, 497.9656145, 511.6400485, 518.8750739, 519.1791709, 514.8609939, 511.4129067, 509.1787248, 510.8071952, 521.4518303, 539.5958691, 558.384097, 570.2413964, 575.9981382, 574.938143, 568.1574012, 565.8462642, 580.1648873, 613.6887855, 648.5767758, 664.3724412, 654.8262788, 631.4170182, 610.8476509, 607.8550885, 627.7045527, 658.6901721, 679.713823, 680.7365818, 674.751457, 676.1155491, 682.9335273, 687.2901818, 694.7895855, 710.2836364, 724.7536873, 729.7160533, 740.0193552, 770.9292606, 810.6269479, 828.4308945, 810.47552, 771.4592582, 741.4206836, 738.2009794, 757.367777, 782.1411297, 799.9823127, 806.4601988, 798.2396509, 783.6566497, 784.2251248, 814.9078885, 863.9618327, 905.5163733, 922.8275588, 916.9566255, 899.9855321, 889.265183, 897.0314473, 918.5466182, 934.683617, 935.6691394, 929.9111564, 930.6682958, 943.5483539, 965.7449503, 990.6709721, 1007.489396, 1006.996635, 990.026783, 968.7002764, 958.0941188, 966.7304727, 986.8070788, 1000.670177, 1004.118264, 1006.239496, 1010.973479, 1013.814613, 1014.383088, 1020.746783, 1036.126642, 1055.217726, 1071.696058, 1083.817736, 1092.605518, 1097.795025, 1097.453692, 1090.445808, 1079.650987, 1066.809406, 1058.43743, 1065.672456, 1091.885615, 1119.197246, 1129.537784, 1127.909314, 1131.659016, 1143.553552, 1157.115035, 1172.304989, 1191.017503, 1203.024989, 1202.343564, 1203.327845, 1217.079234, 1228.632436, 1214.729619, 1178.138686, 1144.57631, 1133.213013, 1144.007835, 1156.053799, 1144.614788, 1111.923743, 1095.029605, 1117.75744, 1151.50848, 1152.682667, 1125.068179, 1111.20384, 1132.227491, 1164.577202, 1173.592126, 1156.887893, 1145.447641, 1170.562327, 1216.699423, 1235.526128, 1202.305086, 1149.690104, 1128.780645, 1149.690104, 1179.540015, 1186.812276, 1173.55489, 1166.092722, 1182.266958, 1211.851248, 1230.60224, 1232.003568, 1229.125198, 1231.700713, 1232.760708, 1227.419772, 1225.034162, 1234.390419, 1247.117808, 1249.731801, 1237.117362, 1216.435045, 1203.51775, 1208.897164, 1222.42017, 1214.275336, 1175.713358, 1139.689658, 1139.273852, 1158.137794, 1156.356655, 1129.613498, 1111.20384, 1113.969261, 1121.28, 1120.825716, 1115.86335, 1102.718914, 1075.104427, 1045.975661, 1035.255312, 1042.300432, 1053.134972, 1057.793241, 1049.535457, 1023.360776, 990.6324945, 970.0258909, 965.4048582, 969.5331297, 983.2075636, 1001.541508, 1002.375602, 977.8666279, 952.6389915, 950.0249988, 962.2223903, 975.4425406, 989.3441164, 996.2762861, 978.3978667, 939.5330327, 904.7592339, 888.9623273, 889.3036606, 903.0538085, 918.8879515, 910.6674036, 868.4699152, 824.3398594, 814.717983, 837.0672485, 856.0069042, 844.6436073, 804.4146812, 760.814623, 738.4653576, 737.9738376, 739.9051636, 735.5869867, 734.7913697, 740.2837333, 737.5940267, 719.6771297, 696.1536776, 679.6008727, 674.6385067, 684.1834279, 702.4044218, 712.7834376, 708.3895467, 699.3733818, 693.8810182, 682.4035297, 661.4183564, 644.6744048, 644.6744048, 653.9177115, 660.1299782, 654.2962812, 628.9544533, 589.1053382, 562.930657, 574.8239515, 608.5377552, 626.6817939, 621.3036218, 612.1745067, 606.9465212, 599.5215903, 591.9464727, 592.2493285, 593.3093236, 591.4909479, 600.0900655, 620.1281939, 627.8559806, 612.211743, 601.4541576, 612.3259345, 625.886177, 624.1062788, 621.0752388, 628.2345503, 633.8038691, 627.7802667, 618.3110594, 609.8248921, 602.3254885, 604.7495758, 619.4467685, 627.401697, 614.4844024, 598.688737, 598.1587394, 603.1968194, 594.7118933, 581.6046933, 583.7644024, 604.4839564, 623.0450424, 617.5911564, 585.8471564, 551.4519273, 541.9070061, 562.9678933, 599.0300703, 633.1212024, 649.5995345, 636.6450036, 601.4541576, 574.2567176, 577.7792776, 605.0909091, 634.6752, 648.4625842, 638.917663, 614.3329745, 598.4615952, 602.3627248, 608.8021333, 603.3097697, 596.8331248, 603.4239612, 615.1285915, 625.810463, 644.3715491, 672.2516558, 692.0254061, 696.0407273, 699.1847176, 708.1239273, 709.6009697, 696.910817, 683.5392388, 680.7750594, 689.2972218, 711.7606788, 746.1931442, 769.9437382, 756.9507297, 717.3659927, 689.4114133, 692.2140703, 711.9878206, 733.8070885, 754.56512, 764.4898521, 750.2841794, 719.6014158, 698.6919564, 699.14624, 714.5633358, 738.3896436, 763.6185212, 770.6636412, 745.0189576, 704.2227976, 682.1763879, 692.3282618, 720.2071273, 744.8290521, 752.5568388, 741.8364897, 723.3895952, 711.6849648, 711.6092509, 719.2228461, 732.9357576, 755.852257, 786.6864485, 810.3613285, 807.7100994, 773.4290618, 732.6316606, 721.9882667, 757.3305406, 814.4523636, 851.9158691, 841.1582836, 791.3459588, 742.8977261, 728.4649115, 738.2009794, 743.0876315, 742.7835345, 758.8448194, 783.9595055, 784.4522667, 756.4207321, 735.8526061, 749.1484703, 782.1796073, 807.44448, 807.5214352, 783.0124606, 751.1939879, 738.8824048, 755.8150206, 786.4220703, 810.9298036, 823.3928145, 823.9228121, 816.0435976, 807.9372412, 803.4676364, 797.6339394, 786.1192145, 776.4216242, 773.4662982, 773.6946812, 775.1332461, 781.6868461, 787.9748267, 778.5043782, 751.7996994, 729.6775758, 734.2228945, 761.3830982, 787.1034958, 795.1726158, 787.8991127, 776.0430545, 767.4054594, 769.1865988, 785.7778812, 813.0510352, 837.8243879, 853.4313891, 862.7876461, 867.0685867, 866.7644897, 865.5530667, 869.41696, 875.2121794, 880.0603539, 888.9250909, 904.8349479, 913.24416, 899.8341042, 871.5754279, 853.1657697, 858.6966109, 884.7583418, 915.7054836, 931.2367709, 915.3653915, 876.8406497, 850.3631127, 861.2348897, 892.7889842, 909.1531248, 902.0695273, 889.9850861, 884.1141527, 878.6217891, 875.9693188, 887.0310012, 908.1676024, 926.3501188, 939.4573188, 947.8665309, 938.7746521, 913.3571103, 908.6603636, 952.03328, 1014.685944, 1044.421663, 1031.050085, 1001.958555, 978.1322473, 963.5864824, 966.1619976, 987.4885042, 1012.716141, 1024.610676, 1019.118313, 1005.292451, 990.6709721, 984.4959418, 985.6328921, 979.4578618, 955.55584, 926.8801164, 915.2512, 922.1076558, 926.7286885, 916.046817, 897.7513503, 884.5312, 874.4165624, 861.1964121, 853.0143418, 863.4318352, 890.7049891, 917.7510012, 932.676577, 933.9637139, 922.0704194, 906.8804655, 909.9487418, 937.714657, 964.0792436, 966.9576145, 956.3129794, 952.3361358, 942.2972121, 908.8117915, 865.9701139, 850.3246352, 871.3867636, 907.0691297, 929.9483927, 928.5855418, 910.1001697, 892.9776485, 893.2432679, 906.0848485, 913.964063, 912.6372073, 909.6074085, 905.0236121, 888.2809018, 863.8488824, 855.098337, 874.6821818, 904.3037091, 916.6525285, 903.8121891, 878.6590255, 858.2423273, 858.3937552, 885.1741479, 926.0087855, 951.5020412, 944.9496824, 911.6144485, 869.8327661, 840.2112388, 845.5149382, 891.9933673, 948.7378618, 968.0176097, 933.8122861, 876.8034133, 837.3328679, 831.2720291, 848.3933091, 865.9316364, 864.7189721, 843.3179927, 821.1201552, 815.9306473, 825.8541382, 834.7188752, 834.3403055, 833.0519273, 838.6957188, 851.3858715, 865.1360194, 878.3561697, 889.6822303, 896.1973527, 895.2887855, 893.0161261, 893.5846012, 899.2283927, 908.4332218, 919.1150933, 921.5019442, 913.5842521, 907.5991273, 916.0852945, 926.3501188, 915.5168194, 883.4699636, 849.4160679, 829.4908897, 833.2033552, 867.8642036, 920.9334691, 951.7304242, 935.6306618, 894.0004073, 870.212577, 875.0235152, 885.5911952, 884.1898667, 876.3863661, 879.0760727, 893.1675539, 904.3421867, 901.2739103, 886.0454788, 867.1443006, 845.9692218, 829.9451733, 836.5744873, 872.5981867, 915.7811976, 934.7220945, 921.6148945, 897.7885867, 886.9925236, 898.8870594, 924.7216485, 945.3282521, 947.5251976, 929.8354424, 904.1150448, 882.4099685, 871.4624776, 868.3942012, 867.8257261, 870.7798109, 881.2357818, 889.6065164, 882.3342545, 864.6060218, 859.4537503, 875.250657, 892.9776485, 894.4559321, 882.1058715, 866.7644897, 853.7727224, 846.537697, 852.4086303, 873.3565673, 902.0695273, 925.1002182, 929.7969648, 911.1229285, 883.3570133, 871.0081939, 885.5154812, 907.5234133, 915.7439612, 910.8573091, 908.3190303, 909.0389333, 900.2883879, 880.1373091, 862.2191709, 858.545183, 872.6366642, 900.2511515, 926.8044024, 931.2367709, 912.5614933, 892.3334594, 882.4099685, 872.143903, 855.3627152, 852.7872, 876.5005576, 907.6748412, 918.8494739, 908.8502691, 896.3115442, 897.9027782, 914.3798691, 931.3509624, 930.0998206, 912.9028267, 906.3877042, 930.1382982, 971.5029333, 999.2688485, 998.7003733, 981.3519515, 971.6543612, 980.8976679, 997.7160921, 1001.958555, 994.2692461, 992.9051539, 1009.686342, 1034.081125, 1044.952902, 1028.815903, 998.0574255, 980.8219539, 995.5948606, 1020.973925, 1023.43649, 996.1633358, 961.6923927, 935.9335176, 926.84288, 943.3956848, 982.5261382, 1017.944126, 1028.853139, 1027.603239, 1030.292945, 1029.72447, 1016.050036, 1006.655302, 1018.361173, 1039.156441, 1044.687282, 1035.633881, 1031.125799, 1042.793193, 1066.089503, 1088.060199, 1093.135515, 1074.423001, 1047.262798, 1037.338065, 1053.929348, 1081.658027, 1102.718914, 1113.249358, 1113.060693, 1103.741673, 1088.060199, 1076.165663, 1072.718817, 1079.006798, 1094.726749, 1111.848029, 1117.568776, 1105.257193, 1085.370492, 1065.521028, 1043.32319, 1020.746783, 1011.617668, 1022.943728, 1037.338065, 1033.626841, 1008.360727, 980.9349042, 972.5256921, 985.7086061, 999.7616097, 990.8224, 962.2608679, 939.3046497, 939.6844606, 957.4499297, 974.4582594, 978.2079612, 968.9646545, 958.2070691, 956.1243152, 956.4271709, 948.7750982, 941.4643588, 946.8810085, 957.8284994, 952.9803248, 931.5023903, 909.001697, 892.6375564, 878.2047418, 866.7644897, 859.9849891, 854.5298618, 848.6589285, 844.8707491, 837.5600097, 820.6646303, 804.1118255, 801.6120242, 804.4903952, 795.1341382, 782.3310352, 788.31616, 812.5582739, 827.8996558, 819.8702545, 796.8383224, 774.6404848, 758.7318691, 750.2841794, 745.9275248, 739.1480242, 724.4123539, 703.4271806, 681.7593406, 664.4866327, 654.107617, 651.4936242, 656.9102739, 663.7667297, 663.084063, 647.0612558, 619.56096, 592.6663758, 574.0283345, 560.3923782, 548.2707006, 542.2470982, 546.7179442, 553.4217309, 544.785377, 518.2693624, 488.6093576, 474.3664485, 481.7913794, 506.0347345, 538.0431127, 561.9078982, 562.4751321, 537.4746376, 503.0036945, 482.0942352, 488.1165964, 515.7310836, 542.8540509, 551.7560242, 543.3840485, 530.7696097, 520.2391661, 511.4129067, 510.1630061, 515.8067976, 515.7695612, 497.6255224, 473.1165479, 456.9038352, 448.9873842, 440.4267442, 437.3584679, 445.2376824, 457.3581188, 461.601823, 458.5323055, 452.8885139, 450.9571879, 455.6166982, 466.7143758, 474.1020703, 470.8438885, 463.3432436, 462.7760097, 471.6010279, 486.6023176, 501.4509382, 506.2618764, 493.0417261, 468.9125624, 447.8504339, 439.8582691, 446.5248194, 466.2600921, 489.7463079, 503.2308364, 501.6408436, 492.1319176, 481.2986182, 470.0867491, 463.4189576, 470.162463, 491.5262061, 518.1936485, 538.3844461, 545.1639467, 541.5284364, 538.231777, 546.6422303, 562.7407515, 573.6125285, 576.0366158, 581.0746958, 597.0217891, 615.734303, 624.0678012, 626.1890327, 638.1592824, 667.8577648, 699.0332897, 710.0937309, 695.5852024, 671.9860364, 662.0613042, 675.5098376, 704.6386036, 728.8447224, 733.4657552, 722.1769309, 712.2521988, 717.1016145, 732.4045188, 744.6776242, 746.7231418, 745.9660024, 754.5266424, 775.4361018, 800.8536436, 820.8172994, 829.5666036, 826.309663, 821.8015806, 828.392417, 849.3775903, 870.9697164, 884.7583418, 893.09184, 900.13696, 902.5610473, 901.7666715, 903.6222836, 908.0161745, 904.68352, 892.0306036, 877.3718885, 867.8257261, 867.105823, 876.2734158, 889.9093721, 895.6301188, 887.5609988, 875.9320824, 874.3793261, 888.5465212, 909.8345503, 921.1233745, 910.1374061, 887.4852848, 878.7347394, 898.2056339, 933.092383, 964.7606691, 985.1016533, 996.2390497, 1002.034269, 1005.405401, 1008.815011, 1016.162987, 1031.884179, 1053.66497, 1070.104824, 1068.210735, 1054.043539, 1041.960339, 1044.421663, 1057.377435, 1070.597585, 1072.528912, 1063.589702, 1054.839156, 1057.641813, 1070.521872, 1082.870691, 1085.63487, 1076.998516, 1062.982749, 1053.66497, 1055.937629, 1066.126739, 1077.756897, 1087.453246, 1094.536844, 1099.650638, 1105.598526, 1115.333353, 1124.538182, 1126.356558, 1119.500102, 1108.628325, 1100.521968, 1101.583205, 1113.060693, 1127.113697, 1131.280446, 1123.93247, 1114.310594, 1108.401183, 1109.423942, 1115.370589, 1124.916752, 1131.129018, 1127.113697, 1112.037935, 1094.991127, 1086.620393, 1090.597236, 1098.817784, 1103.855864, 1104.53729, 1100.939016, 1091.355617, 1079.158225, 1072.226056, 1069.119302, 1061.354279, 1050.293838, 1045.482899, 1045.937183, 1039.724916, 1029.043045, 1028.513047, 1046.088611, 1070.293488, 1087.831816, 1093.855418, 1089.915811, 1080.825173, 1070.861964, 1064.156936, 1064.156936, 1074.195859, 1090.445808, 1099.953493, 1091.014284, 1067.112262, 1045.444422, 1040.671961, 1049.801076, 1058.551622, 1057.036102, 1049.838313, 1041.619006, 1028.777425, 1011.239098, 1002.299888, 1009.193581, 1018.436887, 1013.928805, 1001.01151, 998.7003733, 1009.232058, 1017.565556, 1007.980916, 980.4049067, 950.3663321, 940.0630303, 959.2683055, 986.0114618, 985.1401309, 947.1081503, 900.7054352, 880.5158788, 891.3876558, 910.5544533, 916.0095806, 904.9863758, 887.1067152, 875.1364655, 870.4397188, 867.2572509, 860.8935564, 858.1281358, 862.5977406, 866.38592, 857.2568048, 840.0970473, 829.7937455, 829.4151758, 830.4379345, 827.1425164, 823.5814788, 819.4532073, 813.5810327, 809.4899976, 812.0282764, 814.6422691, 806.6873406, 792.9744291, 788.1647321, 798.1267006, 809.4527612, 810.2483782, 802.7862109, 794.565663, 786.80064, 773.921823, 759.2618667, 749.6027539, 749.2241842, 758.1249164, 777.2172412, 803.9603976, 826.2711855, 834.7188752, 832.1048824, 828.7337503, 829.2265115, 833.2790691, 841.8024727, 850.7416824, 853.4686255, 849.3775903, 847.6361697, 852.9771055, 861.9547927, 869.0383903, 872.9779976, 871.0839079, 863.3933576, 860.0979394, 870.4397188, 885.326817, 884.2655806, 861.2721261, 834.6431612, 825.4755685, 838.9985745, 865.2117333, 888.6222352, 895.6673552, 884.2655806, 870.5911467, 873.0152339, 893.3946958, 914.6827248, 923.5474618, 920.8192776, 911.6901624, 901.0082909, 892.9404121, 893.5846012, 903.2052364, 919.8722327, 940.8201697, 961.8438206, 974.0784485, 972.4884558, 963.2836267, 957.7900218, 962.8665794, 980.5948121, 1006.012354, 1026.58048, 1030.558565, 1020.102594, 1012.451762, 1019.042599, 1037.413779, 1059.271525, 1078.021275, 1090.748664, 1092.756945, 1084.726303, 1072.718817, 1066.240931, 1072.453198, 1090.18143, 1107.454138, 1109.992417, 1096.317983, 1080.066793, 1074.877285, 1081.089552, 1087.984485, 1088.248863, 1082.907927, 1082.226502, 1094.158274, 1112.454982, 1121.735525, 1114.423544, 1099.386259, 1088.476005, 1085.446206, 1090.824378, 1103.211675, 1114.727641, 1112.71936, 1094.688272, 1072.491675, 1061.051423, 1066.846642, 1086.203345, 1106.469857, 1116.697445, 1114.120688, 1107.378424, 1098.855021, 1087.908771, 1073.134623, 1063.399796, 1064.801125, 1075.711379, 1081.99936, 1072.187578, 1051.695166, 1042.263195, 1053.513542, 1069.157779, 1069.233493, 1055.861915, 1047.717081, 1053.626492, 1065.786647, 1074.120145, 1074.005954, 1073.627384, 1084.764781, 1107.529852, 1122.68257, 1113.363549, 1090.332858, 1081.430885, 1098.173595, 1124.386754, 1141.205178, 1145.296213, 1147.720301, 1153.024, 1155.865135, 1150.220102, 1139.122424, 1130.069023, 1125.333799, 1118.477343, 1105.862904, 1091.468567, 1084.764781, 1089.764383, 1099.423496, 1103.250153, 1097.795025, 1086.961726, 1074.650143, 1064.498269, 1063.778366, 1075.711379, 1091.961328, 1098.893498, 1092.869896, 1081.128029, 1077.07423, 1090.56, 1118.174487, 1138.515472, 1133.780247, 1111.848029, 1098.136359, 1101.431777, 1108.325469, 1106.469857, 1101.431777, 1104.15872, 1115.674686, 1122.454187, 1113.173644, 1091.507045, 1075.748616, 1076.998516, 1086.279059, 1089.234385, 1083.513639, 1073.475956, 1064.081222, 1057.377435, 1056.050579, 1057.453149, 1054.952107, 1049.990982, 1044.990138, 1036.315307, 1021.315258, 1011.390526, 1021.808019, 1048.285556, 1066.013789, 1061.278565, 1043.777474, 1035.368262, 1044.157285, 1061.505707, 1069.763491, 1058.854478, 1031.23999, 1002.488553, 987.2613624, 991.1624921, 1004.989595, 1015.481561, 1011.08767, 995.6705745, 981.8819491, 979.3064339, 980.6705261, 974.1541624, 959.4569697, 948.3208145, 946.7680582, 949.9877624, 950.7449018, 944.1155879, 929.2669673, 909.9487418, 896.5771636, 895.1745939, 899.9482958, 902.2966691, 900.932577, 902.2581915, 906.1605624, 901.5382885, 882.5986327, 859.3023224, 853.0528194, 866.3102061, 880.8559709, 878.9618812, 866.8786812, 857.7495661, 853.2414836, 842.2567564, 821.4602473, 799.9065988, 789.603297, 793.0886206, 795.6641358, 782.0654158, 757.6706327, 746.7988558, 760.6631952, 782.0281794, 787.9363491, 776.9143855, 767.1410812, 774.9818182, 801.6120242, 829.4536533, 839.18848, 831.4991709, 830.62784, 848.2791176, 867.5985842, 863.2046933, 840.0598109, 820.4759661, 816.5735952, 815.8921697, 809.1114279, 801.3848824, 802.710497, 811.1569455, 816.5735952, 814.8694109, 810.5897115, 810.3998061, 816.1577891, 825.2869042, 829.4151758, 824.869857, 820.7788218, 831.0064097, 852.5985358, 862.4847903, 848.4317867, 821.8015806, 803.0890667, 797.8610812, 799.9823127, 805.1345842, 812.0667539, 820.0589188, 823.0899588, 814.3009358, 794.7928048, 778.73152, 782.7096048, 803.5433503, 821.2331055, 822.4072921, 814.6422691, 809.6426667, 806.7630545, 796.9525139, 780.7398012, 770.6264048, 778.7699976, 803.6947782, 828.0523248, 832.0291685, 810.3998061, 782.1796073, 772.0277333, 783.2396024, 792.4071952, 779.5656145, 754.2237867, 738.9581188, 736.7226958, 729.7917673, 711.5707733, 697.0634861, 705.2455564, 733.9957527, 767.9739345, 789.8689164, 789.6790109, 770.8150691, 749.3756121, 746.6859055, 767.443937, 795.8155636, 810.5512339, 806.3844848, 794.6798545, 791.9516703, 804.6803006, 826.5740412, 842.0296145, 840.5525721, 826.6497552, 816.1193115, 816.8776921, 820.9687273, 816.0435976, 799.9450764, 780.9297067, 767.3297455, 761.8373818, 763.2387103, 765.3984194, 760.1704339, 741.6478255, 716.5331394, 698.0477673, 698.6534788, 716.1545697, 738.5038352, 749.9813236, 747.3300945, 739.7152582, 741.2307782, 752.3681745, 758.3148218, 745.8902885, 718.5414206, 692.7825455, 683.1991467, 696.1921552, 726.4578715, 755.9279709, 764.5655661, 748.9958012, 726.0035879, 711.457823, 708.8438303, 713.5033406, 720.6998885, 723.2754036, 716.9501867, 704.5256533, 694.10816, 687.9716073, 685.0932364, 684.4105697, 684.9790448, 683.8048582, 676.3042133, 659.2214109, 636.4166206, 618.3867733, 612.6287903, 611.6817455, 603.0826279, 590.6208582, 588.6510545, 604.7495758, 629.2200727, 652.4779055, 669.1833794, 678.7295418, 685.2446642, 695.1309188, 710.0552533, 725.8136824, 739.0723103, 747.9730424, 746.9887612, 735.2456533, 720.018463, 708.3895467, 697.1764364, 682.4035297, 664.675297, 647.5912533, 633.0082521, 629.2200727, 638.7277576, 650.1295321, 645.0914521, 625.9618909, 611.7946958, 616.52992, 630.9627345, 637.9706182, 628.0831224, 609.0305164, 592.5137067, 581.8703127, 572.2112, 565.2405527, 567.0589285, 574.7110012, 573.2711952, 558.4598109, 544.8610909, 544.6724267, 553.9902061, 563.7250327, 574.9753794, 589.7495273, 601.1885382, 600.9998739, 587.7040097, 569.7101576, 557.2099103, 559.1797139, 572.5140558, 581.0746958, 570.7713939, 547.9293673, 531.8680824, 531.8308461, 542.0956703, 556.7556267, 573.3853867, 588.0441018, 592.9692315, 585.8856339, 571.1884412, 553.4602085, 538.8387297, 533.6492218, 539.6343467, 547.0580364, 548.4593648, 550.4291685, 566.3017891, 597.5145503, 625.2804655, 634.6752, 630.2043539, 627.5158885, 632.2883491, 634.7881503, 625.5076073, 602.4012024, 572.3241503, 548.3836509, 545.6951855, 560.3923782, 572.1354861, 565.4304582, 553.4602085, 557.1341964, 581.7945988, 617.0599176, 652.0620994, 675.3571685, 680.2065842, 668.9934739, 654.523423, 646.7584, 649.9023903, 666.7592921, 692.896737, 715.8517139, 729.5261479, 739.8294497, 747.3300945, 735.4727952, 696.7978667, 650.4708655, 626.5688436, 629.2957867, 638.1592824, 636.3409067, 623.6892315, 610.280417, 601.9084412, 603.3097697, 620.2051491, 652.8564752, 692.896737, 726.7607273, 744.2233406, 738.655263, 712.9721018, 685.4718061, 679.4866812, 703.5786085, 740.3979248, 765.9668945, 766.1555588, 745.700383, 722.9725479, 719.4115103, 742.5191564, 777.6330473, 801.5363103, 803.5818279, 790.8531976, 779.7927564, 782.6711273, 800.2479321, 822.55872, 836.1959176, 834.6803976, 821.1201552, 807.6343855, 804.9074424, 812.9380848, 822.4072921, 826.0440436, 825.3998545, 825.4755685, 828.6195588, 835.8173479, 844.7193212, 852.4843442, 860.6664145, 875.7049406, 895.0616436, 907.0318933, 909.1903612, 912.7886352, 925.7059297, 937.5247515, 938.8131297, 933.092383, 930.9711515, 933.5094303, 935.5934255, 934.3050473, 933.6608582, 940.6687418, 955.9344097, 966.8061867, 960.6696339, 939.5702691, 919.9479467, 911.7658764, 909.1531248, 905.9334206, 899.1899152, 890.7434667, 882.295777, 879.98464, 886.2353842, 894.7587879, 900.4398158, 906.5391321, 914.3798691, 915.4026279, 907.3347491, 903.6222836, 919.0393794, 945.5169164, 964.6092412, 967.7147539, 960.8582982, 955.2145067, 960.1396364, 976.5410133, 991.9965867, 995.1778133, 984.2688, 967.8289455, 949.1536679, 936.9947539, 941.237217, 959.798303, 972.7155976, 964.9493333, 942.7142594, 921.6906085, 914.455583, 929.3426812, 958.8127806, 978.5865309, 973.397023, 958.7370667, 959.0026861, 975.9725382, 993.5108655, 1004.723976, 1011.011956, 1015.557275, 1019.004121, 1020.330977, 1017.792698, 1016.997081, 1032.565605, 1063.172655, 1087.908771, 1095.029605, 1100.029207, 1118.060296, 1139.198138, 1148.288776, 1144.917644, 1139.235375, 1138.666899, 1146.129067, 1156.092276, 1151.811336, 1127.189411, 1098.552165, 1083.62783, 1079.195462, 1077.794133, 1086.393251, 1110.863748, 1135.447195, 1140.145183, 1125.295321, 1107.112805, 1099.272068, 1107.264233, 1127.227888, 1140.333847, 1128.780645, 1102.074725, 1091.468567, 1111.432223, 1143.970599, 1164.350061, 1168.592524, 1163.820063
-};
-
-float realNoise2T[255] {
-  621.1351584,595.7085736,571.4927779,546.0661924,521.8503968,496.4238113,472.2080156,446.7814302,422.5656345,398.3498388,372.9232534,348.7074577,323.2808722,299.0650766,273.6384911,249.4226954,223.99611,199.7803143,175.5645186,150.1379332,125.9221375,100.495552,76.27975637,50.85317091,26.63737524,615.0812095,590.8654144,567.8604085,543.6446129,520.639607,496.4238113,473.4188054,449.2030097,426.1980038,401.9822076,378.9772023,354.7614066,331.7564007,308.7513943,284.5355992,261.5305933,237.3147976,214.3097911,190.093996,167.0889901,142.8731945,119.868188,95.65239291,72.64738702,48.43159135,25.42658546,615.0812095,590.8654144,566.6496187,543.6446123,519.4288172,495.2130215,472.2080156,447.9922194,424.9872141,400.7714184,376.5556227,353.5506162,329.3348212,306.3298153,282.1140196,257.8982233,234.893218,210.6774224,186.4616261,163.4566202,139.2408251,116.2358192,92.02002356,67.80422731,44.79922199,20.58342632,615.0812095,590.8654138,566.6496187,543.6446129,519.4288172,496.4238107,472.2080156,449.2030097,424.9872135,401.9822076,377.7664125,354.7614066,330.5456109,307.5406045,283.3248094,259.1090137,236.1040072,211.8882116,188.8832062,164.6674106,141.6624047,117.4466084,94.44160312,70.22580745,47.22080098,23.00500589,612.6596299,588.4438349,565.438829,541.2230327,518.2180274,494.0022317,470.9972253,446.7814296,423.7764243,399.5606286,376.5556227,352.3398265,329.3348212,306.3298153,282.114019,259.1090131,234.893218,211.8882121,187.6724165,164.66741,140.4516149,117.446609,93.23081276,70.22580687,46.01001178,23.00500589,618.7135789,594.4977838,570.2819881,547.2769816,523.061186,498.8453903,475.840385,451.6245893,427.4087936,404.4037877,380.1879915,355.9721958,332.9671899,308.7513943,284.5355992,261.5305933,237.3147976,213.0990019,190.093996,165.8781998,141.6624041,118.6573982,94.44160312,70.22580745,47.22080156,23.00500589,615.0812095,590.8654138,567.860408,543.6446123,520.639607,497.6346011,473.4188054,450.4137995,427.4087936,403.1929974,380.1879915,357.1829856,332.9671905,309.9621846,286.9571787,262.7413831,239.7363766,216.7313707,192.515575,169.5105691,146.5055638,122.2897681,99.28476226,76.27975637,52.0639607,29.05895481,621.1351584,596.9193628,572.7035677,549.6985618,525.4827661,502.4777602,478.261964,455.2569581,431.0411624,408.0361565,383.8203614,360.8153555,336.5995599,313.594554,289.3787583,265.162962,242.1579562,217.9421605,194.9371552,170.7213595,147.7163536,123.5005579,100.4955515,76.27975579,53.2747499,29.05895481,615.0812095,590.8654144,566.6496187,543.6446129,519.4288172,496.4238107,472.208015,449.2030092,424.9872141,401.9822082,377.7664125,354.7614066,330.5456104,307.5406045,283.3248088,259.1090131,236.1040078,211.8882121,188.8832062,164.6674106,141.6624047,117.4466084,94.44160255,70.22580687,47.22080156,23.00500589,615.0812095,584.8114655,555.7525101,526.6935553,497.6346005,468.5756457,438.3059017,409.2469469,380.1879921,351.1290373,322.0700824,291.8003373,262.7413825,233.6824277,204.6234734,175.5645186,145.294774,116.2358192,87.17686385,58.11790904,29.05895423,0
-};
-
-float realNoise4T2[1729] {
-  1156.697988, 1149.765818, 1148.36449, 1156.280941, 1171.357944, 1182.342672, 1173.062128, 1139.53823, 1099.613401, 1078.096989, 1083.286497, 1102.946056, 1121.204286, 1125.977988, 1115.7504, 1099.840543, 1090.218667, 1085.256301, 1074.537193, 1058.173052, 1041.504815, 1022.452208, 1000.215893, 986.4657455, 989.8368776, 998.0189479, 993.2464873, 977.4123442, 963.9265745, 965.1020024, 989.4583079, 1033.058366, 1073.097387, 1089.498764, 1095.332461, 1112.302313, 1133.023108, 1131.24321, 1110.976698, 1102.908819, 1120.105813, 1139.955278, 1138.666899, 1113.439263, 1077.756897, 1045.975661, 1028.247428, 1020.746783, 1009.383486, 986.6928873, 961.9952485, 947.2223418, 939.0402715, 932.600863, 929.2297309, 928.3584, 915.5168194, 881.4629236, 844.5678933, 833.0904048, 855.9311903, 896.0459248, 932.676577, 953.6989867, 959.7598255, 961.6166788, 972.8285479, 996.0491442, 1024.800582, 1055.217726, 1084.0064, 1100.333304, 1092.642754, 1065.823884, 1038.512252, 1026.58048, 1027.300383, 1024.042201, 1004.345406, 972.0714085, 942.0700703, 918.7365236, 900.5540073, 895.1745939, 908.2817939, 924.7588848, 922.600417, 903.6607612, 889.0380412, 890.3264194, 906.7662739, 938.0547491, 979.4963394, 1011.883287, 1022.299539, 1017.64127, 1010.973479, 1009.723578, 1019.53536, 1041.050531, 1057.338958, 1045.255758, 1006.466638, 967.4118982, 950.6307103, 954.5330812, 963.1321988, 962.1466764, 946.9567224, 924.607457, 906.5763685, 898.2813479, 892.5990788, 882.0301576, 868.3557236, 863.1289794, 867.6742982, 872.7880921, 872.143903, 871.5754279, 876.1592242, 878.3189333, 868.5828655, 845.5521745, 816.309217, 793.6943321, 792.6715733, 813.99808, 831.5748848, 818.3919709, 778.3529503, 741.9122036, 732.5187103, 748.1629479, 770.1336436, 781.2325624, 781.6111321, 785.5122618, 800.5507879, 816.0820752, 824.5285236, 827.1425164, 821.8015806, 800.0580267, 772.0649697, 761.2689067, 770.6264048, 773.8076315, 752.4811248, 712.2149624, 671.9103224, 645.3943079, 642.4774594, 660.9255952, 690.9269333, 723.2381673, 753.5038836, 771.8005915, 774.9445818, 776.1560048, 790.4373915, 809.0741915, 815.8549333, 807.1801018, 792.1030982, 777.7844752, 772.2548752, 778.1258085, 783.0124606, 769.1481212, 743.9577212, 730.7003345, 735.8898424, 738.8439273, 725.2079709, 703.9944145, 693.5024485, 702.2902303, 724.0337842, 742.7835345, 744.6019103, 730.170337, 711.6464873, 700.2074764, 696.3808194, 699.4118594, 703.2000388, 698.7291927, 680.0166788, 657.1374158, 648.3111564, 661.9483539, 692.4039758, 724.4495903, 739.5265939, 727.063583, 698.8433842, 682.4792436, 690.1313164, 703.5401309, 700.6989964, 689.1842715, 693.1611152, 722.2154085, 760.4745309, 789.7932024, 809.0357139, 824.7184291, 837.8243879, 837.4842958, 822.862817, 811.4982788, 826.309663, 863.7346909, 889.0765188, 866.4616339, 802.1035442, 738.5410715, 714.4876218, 724.1094982, 731.5716655, 717.7458036, 696.910817, 688.8429382, 690.9641697, 691.0026473, 691.1155976, 703.3514667, 729.7532897, 761.4588121, 791.9516703, 825.3626182, 855.3627152, 866.0830642, 842.5596121, 796.0427055, 754.7922618, 739.4881164, 746.3445721, 754.56512, 744.8675297, 712.2906764, 673.6157479, 653.5763782, 658.3873164, 670.5474715, 673.95584, 674.1457455, 682.2893382, 697.8206255, 713.8434327, 721.3813139, 710.4350642, 679.0323976, 650.8866715, 654.0691394, 687.2517042, 717.9729455, 720.9270303, 695.1309188, 656.8717964, 629.0686448, 631.3785406, 663.7667297, 702.2902303, 720.7383661, 712.1764848, 687.8201794, 665.2437721, 660.9255952, 683.0477188, 721.8368388, 756.2693042, 770.7393552, 770.7393552, 774.8688679, 796.081183, 827.6352776, 852.4086303, 861.9547927, 861.5377455, 860.0222255, 860.627937, 863.0905018, 862.2948848, 849.5674958, 818.6948267, 778.4286642, 751.6482715, 757.2535855, 791.9144339, 833.0519273, 859.1893721, 866.8414448, 865.7417309, 866.8786812, 872.4852364, 881.4244461, 894.60736, 913.3198739, 929.2297309, 925.3658376, 898.3185842, 873.4695176, 878.3189333, 908.9632194, 938.43456, 948.7378618, 938.92608, 914.6070109, 890.9706085, 890.6292752, 913.2056824, 926.4258327, 908.7745552, 875.9320824, 853.4686255, 846.9162667, 854.4156703, 875.5535127, 900.5924848, 915.1754861, 925.4030739, 952.9046109, 1000.178657, 1044.687282, 1070.066347, 1082.642308, 1095.143796, 1105.901382, 1098.439215, 1063.209891, 1012.413285, 971.6171248, 951.2748994, 945.6695855, 948.4337648, 957.3357382, 966.2377115, 974.1541624, 989.2311661, 1011.201862, 1022.110875, 1017.981362, 1023.739345, 1055.67201, 1087.188868, 1079.158225, 1029.194473, 973.6998788, 945.3654885, 948.9650036, 967.8289455, 983.0188994, 987.4127903, 980.2150012, 963.0180073, 939.5702691, 920.8949915, 917.4109091, 926.0087855, 931.0096291, 921.3120388, 900.7054352, 883.9242473, 876.8406497, 874.94656, 870.5911467, 861.8406012, 848.5075006, 832.0291685, 822.0287224, 826.6497552, 837.9001018, 840.6282861, 832.1805964, 817.02912, 791.3844364, 758.3905358, 744.7161018, 767.1026036, 802.2549721, 810.7411394, 790.5888194, 772.8990642, 780.24704, 806.0431515, 825.8541382, 821.5359612, 793.8842376, 764.6028024, 754.9052121, 769.2623127, 794.9442327, 816.8019782, 821.347297, 809.9455224, 789.8689164, 764.1857552, 730.4731927, 704.0701285, 708.3138327, 741.5721115, 772.5192533, 778.3529503, 759.9817697, 728.654817, 699.6005236, 691.1913115, 706.8355491, 723.2754036, 722.8595976, 716.3059976, 721.9125527, 739.1480242, 757.2535855, 769.0351709, 771.1179248, 768.3525042, 772.2548752, 782.5196994, 780.1713261, 752.0268412, 713.6547685, 687.3658958, 679.7523006, 682.4792436, 687.9343709, 691.1913115, 688.3129406, 683.3505745, 683.5020024, 692.9724509, 707.1768824, 719.9799855, 728.9961503, 740.8522085, 759.9817697, 781.042657, 791.83872, 790.5503418, 782.0281794, 770.2093576, 757.8977745, 752.0268412, 760.2461479, 778.0128582, 795.2855661, 801.6120242, 796.0427055, 788.3918739, 796.7626085, 829.4536533, 874.7578958, 910.8957867, 923.32032, 908.8117915, 878.8861673, 856.5753794, 861.4620315, 884.3412945, 895.0231661, 879.0760727, 855.7797624, 851.7272048, 867.3714424, 881.9172073, 881.0831127, 866.8786812, 848.9990206, 835.9302982, 838.392863, 858.0151855, 879.2659782, 884.9854836, 878.3189333, 876.1977018, 876.3491297, 864.4545939, 842.71104, 841.19552, 877.2204606, 936.7291345, 984.7987976, 1001.76865, 988.4355491, 960.2910642, 932.2595297, 914.9111079, 914.039777, 932.2210521, 961.9952485, 983.7760388, 981.8062352, 958.7755442, 937.1077042, 934.1151418, 951.8818521, 976.1624436, 993.2080097, 1000.784368, 1011.883287, 1039.307869, 1073.210337, 1086.506201, 1066.543787, 1034.496931, 1021.42945, 1031.845702, 1046.391467, 1055.483345, 1070.104824, 1095.068082, 1118.325915, 1126.053702, 1122.302759, 1119.38591, 1124.083898, 1136.091384, 1149.690104, 1152.304097, 1136.357004, 1116.014778, 1118.855913, 1147.000398, 1166.281387, 1147.910206, 1108.893944, 1091.999806, 1117.37887, 1170.448136, 1227.685392, 1261.966429, 1247.269236, 1184.691045, 1120.939908, 1112.643646, 1165.297105, 1233.443375, 1265.034705, 1245.867908, 1202.116422, 1178.70592, 1199.616621, 1243.898104, 1271.134022, 1263.784805, 1239.087166, 1218.026279, 1208.177261, 1202.343564, 1188.366274, 1165.031486, 1149.008679, 1161.925973, 1197.495389, 1225.261304, 1224.502924, 1204.503273, 1192.608737, 1210.108587, 1253.89855, 1296.21023, 1302.422497, 1265.034705, 1217.11647, 1204.691937, 1242.079728, 1291.323578, 1301.24831, 1264.087661, 1223.367215, 1217.912087, 1240.564208, 1261.285004, 1266.967273, 1258.973867, 1240.754114, 1221.358933, 1217.533518, 1228.593959, 1237.003171, 1233.026327, 1226.359777, 1220.14751, 1205.336126, 1181.774196, 1171.243753, 1184.275239, 1200.903758, 1193.895874, 1165.978531, 1146.660305, 1152.569716, 1166.508528, 1160.107598, 1126.393794, 1089.498764, 1080.635268, 1110.598128, 1151.622672, 1161.357498, 1129.046264, 1089.234385, 1081.696504, 1102.188916, 1112.984979, 1088.778861, 1040.55777, 1003.28417, 1005.253973, 1038.512252, 1069.839205, 1071.391961, 1054.687728, 1040.217678, 1031.694274, 1018.285459, 1006.201018, 1012.867568, 1041.543292, 1071.506153, 1078.665464, 1052.869353, 999.9130376, 935.7063758, 882.0301576, 860.2121309, 873.5452315, 898.7356315, 902.8266667, 877.3334109, 842.2567564, 819.4147297, 811.6497067, 806.6873406, 793.2785261, 771.8378279, 756.0793988, 761.3830982, 779.8684703, 781.5354182, 744.5646739, 682.2521018, 631.64416, 623.5005673, 657.7816048, 704.6013673, 721.9497891, 687.3658958, 618.3867733, 559.1424776, 543.6869042, 561.5665648, 575.0895709, 559.4453333, 527.3984776, 503.913503, 500.1253236, 511.2614788, 519.8605964, 506.1861624, 469.3283685, 442.3580703, 452.1313745, 477.0561552, 477.3590109, 447.5103418, 415.4634861, 400.6148655, 405.46304, 423.5698424, 434.3274279, 414.5164412, 372.4691394, 343.6804655, 345.4988412, 362.355743, 377.6971248, 386.3719564, 385.0463418, 371.4463806, 351.5981576, 329.3246061, 309.8537115, 307.6195297, 331.5215515, 362.1658376, 371.9391418, 360.6515588, 346.0673164, 336.1425842, 332.3171685, 337.96096, 352.6209164, 370.8406691, 393.6442182, 419.3646158, 432.5090521, 421.9786085, 403.456, 399.5163927, 416.0319612, 446.6762473, 476.2605382, 481.6771879, 452.5856582, 414.7820606, 400.42496, 409.4026473, 427.0551661, 457.9650715, 507.2846352, 553.3844945, 569.8243491, 561.1495176, 547.8921309, 542.6641455, 553.3087806, 588.3097212, 635.8493867, 665.6980558, 662.9326352, 646.2284024, 641.0761309, 652.3264776, 669.7518545, 683.8433358, 690.3956945, 692.5554036, 696.7978667, 705.2070788, 707.70688, 693.0854012, 662.9326352, 632.1741576, 616.6056339, 628.841503, 673.5015564, 735.6627006, 782.8225552, 788.8833939, 759.9817697, 728.1248194, 721.9882667, 746.7231418, 783.6938861, 810.5127564, 818.6191127, 815.9306473, 812.3311321, 812.0282764, 813.6195103, 818.1648291, 827.5210861, 848.6204509, 881.0073988, 907.5991273, 901.9180994, 864.6432582, 829.8694594, 831.5748848, 867.8257261, 912.0699733, 939.8358885, 948.5094788, 950.4792824, 961.3510594, 977.2981527, 983.0561358, 979.0035782, 989.4583079, 1026.126196, 1068.552068, 1090.294381, 1097.265028, 1110.900984, 1139.689658, 1169.842424, 1189.350555, 1198.706812, 1207.722977, 1222.53312, 1237.723074, 1240.375544, 1228.253867, 1221.397411, 1243.443821, 1291.247864, 1330.605459, 1333.332402, 1306.892102, 1279.618948, 1268.860121, 1269.618502, 1269.239932, 1262.193571, 1251.321794, 1248.594851, 1260.792242, 1280.830371, 1291.853576, 1289.959486, 1288.861013, 1299.278507, 1318.596732, 1329.468509, 1319.355113, 1293.330618, 1273.330967, 1274.883724, 1288.255302, 1290.490725, 1274.618104, 1256.701207, 1255.109973, 1273.898201, 1295.45185, 1293.974807, 1265.564703, 1240.222875, 1252.64865, 1295.680233, 1331.059743, 1332.536785, 1307.384863, 1277.042192, 1262.723568, 1273.103825, 1296.020325, 1304.580965, 1289.126633, 1268.898599, 1264.73185, 1274.163821, 1286.058356, 1298.48289, 1309.240475, 1309.543331, 1294.657474, 1279.087709, 1277.080669, 1286.89121, 1291.92929, 1279.011995, 1251.057416, 1228.935292, 1235.64032, 1269.76993, 1300.794027, 1298.444412, 1269.239932, 1244.162482, 1242.382584, 1250.261799, 1241.245634, 1206.396121, 1165.221392, 1145.826211, 1149.99296, 1149.50144, 1120.825716, 1074.005954, 1040.55777, 1037.79359, 1058.096097, 1072.870245, 1055.066298, 1003.398361, 952.5632776, 935.9719952, 956.2372655, 986.4657455, 1002.715695, 989.7996412, 952.5632776, 915.3653915, 897.9027782, 886.2353842, 861.34784, 840.4011442, 849.0374982, 871.1968582, 864.3404024, 825.9683297, 790.8531976, 783.0881745, 797.672417, 819.2633018, 832.9377358, 826.9153745, 803.0133527, 773.7319176, 741.0421139, 705.0171733, 677.8197333, 677.3269721, 701.7217552, 730.4347152, 741.6850618, 726.7234909, 693.9567321, 667.1378618, 666.0766255, 688.9943661, 718.0858958, 730.3974788, 711.4193455, 666.9479564, 626.9846497, 617.325537, 631.64416, 644.6744048, 638.4621382, 610.0532752, 568.7258764, 533.8763636, 522.5875394, 524.1787733, 514.8609939, 492.0946812, 474.2150206, 471.9051248, 480.0871952, 491.6403976, 499.0640873, 490.6176388, 463.3060073, 434.8201891, 424.8582206, 435.7287564, 448.6845285, 445.8049164, 420.0088048, 384.099297, 361.5228897, 369.5150545, 397.8866812, 416.6004364, 403.3045721, 369.8179103, 350.9154909, 365.2341139, 396.3339248, 414.0993939, 404.5159952, 377.3185552, 355.1579539, 358.2647079, 388.1903321, 422.6227976, 436.9414206, 427.6981139, 410.0840727, 400.5006739, 410.9181673, 446.979103, 497.8141867, 535.1635006, 537.5131152, 513.1183321, 491.1091588, 493.6102012, 517.4365091, 540.0886303, 545.0125188, 533.3463661, 520.2776436, 515.7695612, 522.2474473, 535.1635006, 549.1420315, 563.1950352, 581.1119321, 601.8712048, 614.0301188, 607.2878545, 588.3854352, 576.3779491, 589.2567661, 625.7347491, 666.5693867, 684.032, 668.8048097, 637.8949042, 614.9386861, 614.1443103, 636.1522424, 672.8958448, 705.2827927, 717.5186618, 712.6692461, 705.1698424, 700.0932848, 697.5934836, 699.7904291, 708.2753552, 713.9576242, 713.4276267, 715.8517139, 727.9349139, 739.5265939, 739.6023079, 735.3598448, 743.4277236, 769.5651685, 800.5135515, 816.9534061, 811.4982788, 796.3840388, 791.5731006, 806.2702933, 833.0904048, 860.7421285, 882.9772024, 898.2813479, 905.5536097, 908.1676024, 913.0927321, 922.4862255, 929.6083006, 930.9711515, 935.8962812, 953.9273697, 978.6622448, 995.0263855, 999.7231321, 998.5117091, 996.1633358, 995.8220024, 1005.443879, 1025.292102, 1042.527573, 1044.421663, 1035.975215, 1029.345901, 1029.497328, 1034.383981, 1041.429101, 1050.293838, 1057.073338, 1058.70305, 1050.369552, 1034.725314, 1019.079835, 1010.594909, 1006.390924, 1004.382642, 1008.549392, 1024.800582, 1044.877188, 1055.823438, 1049.990982, 1037.906541, 1041.201959, 1074.384524, 1116.470303, 1128.439312, 1094.991127, 1053.967825, 1049.081173, 1082.49088, 1119.234482, 1135.258531, 1134.575864, 1131.318924, 1129.766167, 1123.098376, 1107.037091, 1092.18847, 1099.537687, 1133.894439, 1169.198235, 1170.675278, 1137.682618, 1098.400737, 1075.218618, 1067.907879, 1074.877285, 1098.74207, 1124.121135, 1121.999903, 1085.067636, 1043.777474, 1029.838662, 1046.808514, 1074.574429, 1091.089998, 1085.219064, 1064.688175, 1050.14241, 1049.422507, 1048.702604, 1035.595404, 1018.095554, 1007.337968, 1004.609784, 1007.035113, 1015.520039, 1025.520485, 1023.966487, 1001.655699, 961.3510594, 919.3807127, 898.0542061, 905.6678012, 922.4489891, 922.9405091, 904.7207564, 886.6896679, 881.2357818, 887.5237624, 899.64544, 911.1986424, 916.8424339, 914.5697745, 906.1220848, 894.2660267, 884.7955782, 882.9772024, 881.8787297, 868.3184873, 844.9836994, 822.3700558, 801.5363103, 777.8614303, 762.5572848, 766.7997479, 778.9214255, 777.6715248, 761.0417648, 739.2609745, 712.5935321, 684.9033309, 675.8126933, 700.6232824, 739.7537358, 760.7773867, 753.9209309, 738.9965964, 729.6390982, 721.1169358, 707.1011685, 698.1234812, 708.3138327, 733.2386133, 750.7769406, 741.1178279, 703.3514667, 660.8498812, 646.8725915, 678.274017, 730.0189091, 757.4062545, 738.0495515, 694.7523491, 665.6608194, 667.1750982, 681.0394376, 684.3348558, 677.4026861, 681.305057, 699.3361455, 711.0780121, 703.7300364, 689.9798885, 682.2893382, 677.5553552, 672.2516558, 673.7671758, 678.8052558, 675.7369794, 664.675297, 658.1986521, 659.7129309, 659.8643588, 656.7588461, 659.1829333, 672.1759418, 686.1532315, 682.9335273, 657.1374158, 627.9316945, 618.1211539, 620.0909576, 609.9006061, 589.5596218, 582.4387879, 589.1425745, 583.4615467, 557.1341964, 530.39104, 523.5345842, 540.0129164, 574.1425261, 612.4016485, 633.5382497, 630.5456873, 620.2051491, 623.4620897, 643.5386958, 667.5921455, 680.0923927, 675.5470739, 662.3654012, 653.2735224, 650.0538182, 650.0538182, 652.8949527, 660.6227394, 670.3575661, 681.2665794, 693.8053042, 696.1921552, 679.5239176, 658.0087467, 660.5842618, 688.2372267, 711.1537261, 708.5782109, 693.388257, 690.9269333, 706.3055515, 722.6324558, 727.7462497, 733.9957527, 759.4132945, 796.1941333, 814.6422691, 804.7560145, 788.6190158, 792.9744291, 816.309217, 840.9683782, 855.2497648, 857.9767079, 854.6055758, 848.9990206, 847.3705503, 853.6970085, 865.6287806, 872.4095224, 865.7044945, 852.6742497, 845.2865552, 846.6134109, 846.6891248, 838.8099103, 823.3928145, 811.3468509, 815.2107442, 836.2343952, 853.2414836, 844.1508461, 815.3994085, 796.0427055, 803.1647806, 826.9153745, 844.0366545, 844.3407515, 840.6282861, 851.6887273, 876.2734158, 892.296223, 891.3876558, 887.4480485, 890.856417, 896.3115442, 902.8651442, 917.7510012, 932.600863, 934.3807612, 933.1308606, 953.3588945, 991.8451588, 1021.201067, 1031.391418, 1037.452257, 1052.036499, 1067.605023, 1073.514434, 1068.854924, 1059.801522, 1054.194967, 1054.839156, 1063.551224, 1079.195462, 1092.718468, 1088.551719, 1059.838759, 1023.550681, 1003.738453, 1007.867966, 1024.307821, 1040.369105, 1050.709644, 1052.906589, 1043.437382, 1025.482007, 1011.08767, 1010.443481, 1021.845256, 1036.732354, 1053.210686, 1066.468073, 1062.339801, 1032.75551, 1000.178657, 990.3296388, 1002.185697, 1015.330133, 1028.133236, 1049.686885, 1069.612063, 1069.006352, 1047.414225, 1029.194473, 1029.497328, 1043.70176, 1057.226007, 1062.794085, 1064.725411, 1062.22561, 1049.498221, 1031.201513, 1028.740189, 1048.664126, 1068.51359, 1072.907481, 1073.248815, 1075.407282, 1064.725411, 1040.293392, 1026.543244, 1027.490288, 1018.285459, 996.1248582, 989.9510691, 1007.56511, 1018.966885, 1004.79969, 985.064417, 985.6701285, 1002.678458, 1016.277178, 1013.398807, 997.4889503, 978.8136727, 963.0180073, 952.3733721, 951.7676606, 962.3738182, 972.3742642, 966.0862836, 944.8739685, 925.5545018, 916.7282424, 916.5395782, 918.0550982, 916.4253867, 906.0848485, 890.4021333, 881.6143515, 885.6296727, 893.1675539, 889.8336582, 875.4393212, 861.7648873, 851.500063, 831.1206012, 790.3232, 748.9585648, 739.7537358, 770.6264048, 811.4225648, 828.8851782, 816.7262642, 789.111777, 760.7016727, 739.6780218, 733.6544194, 743.9204848, 760.814623, 769.7923103, 770.5879273, 769.5651685, 758.1249164, 724.4123539, 688.3129406, 689.9798885, 727.9349139, 749.6399903, 721.8368388, 673.0472727, 642.4017455, 638.7662352, 652.4406691, 678.6910642, 703.4271806, 713.1620073, 713.0862933, 718.2373236, 722.7454061, 713.4276267, 698.0105309, 692.8197818, 697.1392, 696.8735806, 689.7142691, 683.236383, 682.4407661, 683.0477188, 678.1610667, 669.4105212, 667.3277673, 675.3956461, 682.7820994, 686.5702788, 698.5020509, 721.5712194, 735.0942255, 723.1239758, 699.6390012, 685.2061867, 681.0022012, 673.7671758, 659.0687418, 645.24288, 643.2730764, 653.5763782, 657.554463, 637.6292848, 604.5981479, 591.2638061, 615.318497, 659.2586473, 694.6381576, 708.6924024, 707.1768824, 700.8504242, 696.419297, 697.2521503, 704.7143176, 717.6316121, 728.7305309, 730.7760485, 722.101217, 706.1541236, 687.2901818, 675.3571685, 680.9264873, 704.3742255, 730.8517624, 744.3375321, 736.1169842, 707.4040242, 673.3501285, 660.8871176, 685.1689503, 730.5489067, 767.2925091, 779.2999952, 769.2995491, 746.7988558, 722.82112, 708.1239273, 704.1470836, 706.4569794, 712.9348655, 727.3664388, 746.7988558, 757.0649212, 747.8216145, 724.9051152, 706.419743, 703.388703, 713.9576242, 727.063583, 733.5787055, 728.2377697, 718.8057988, 719.3357964, 737.93536, 764.9056582, 787.8233988, 804.641823, 817.8619733, 826.0825212, 830.4006982, 837.67296, 847.7875976, 847.6734061, 828.8851782, 804.4903952, 797.672417, 815.5508364, 840.5140945, 848.544737, 838.3171491, 831.0821236, 841.8397091, 862.4463127, 874.7951321, 868.621343, 845.4764606, 822.1801503, 825.4755685, 861.4992679, 893.2047903, 886.6139539, 856.2725236, 847.029217, 865.5145891, 883.9242473, 884.9470061, 882.9014885, 886.7281455, 893.5846012, 898.0542061, 900.1754376, 896.8800194, 889.3036606, 892.4104145, 918.6980461, 954.1917479, 970.7457939, 958.6613527, 939.1532218, 933.39648, 936.4262788, 933.2438109, 923.6988897, 917.0323394, 911.7658764, 902.1452412, 893.7732655, 896.5771636, 905.8577067, 912.4857794, 910.3645479, 896.8800194, 875.4777988, 863.4690715, 880.9316848, 917.6752873, 945.6311079, 952.0705164, 946.5396752, 934.9492364, 916.4253867, 897.410017, 889.3036606, 891.3491782, 898.0157285, 908.7745552, 921.5391806, 924.4945067, 915.8196752, 907.4476994, 908.7745552, 912.524257, 915.5540558, 927.3728776, 951.8061382, 969.4201794, 960.5182061, 931.9566739, 912.4100655, 918.6608097, 939.2289358, 952.5632776, 950.3663321, 937.5632291, 923.4717479, 916.9566255, 925.4030739, 942.2599758, 948.5479564, 936.0849455, 919.7965188, 917.3724315, 924.7588848, 932.7522909, 940.5930279, 945.1768242, 935.1379006, 916.2367224, 914.6827248, 940.5557915, 962.0709624, 951.9948024, 919.8722327, 894.60736, 884.4170085, 886.6139539, 905.5536097, 935.3662836, 952.7147055, 939.3803636, 909.417503, 887.7509042, 882.0686352, 887.2196655, 900.932577, 924.1916509, 948.1309091, 961.5409648, 959.874017, 946.5396752, 926.0087855, 906.4249406, 896.2358303, 899.0384873, 908.0918885, 912.2214012, 906.3877042, 897.0314473, 894.0388848, 899.7968679, 903.2052364, 893.0161261, 873.8108509, 862.2948848, 865.9316364, 874.2651345, 871.3110497, 858.620897, 852.2187248, 864.1889745, 885.1741479, 894.7960242, 887.333857, 876.3863661, 874.8336097, 876.6892218, 869.0383903, 848.4317867, 823.2029091, 807.2558158, 813.4680824, 839.6427636, 864.6060218, 867.5985842, 854.0755782, 844.0366545, 844.6436073, 848.6961648, 850.0217794, 849.0747345, 842.2567564, 828.0523248, 813.0895127, 807.1801018, 812.0667539, 821.5744388, 826.0825212, 820.2103467, 807.3315297, 797.672417, 798.88384, 806.9144824, 813.0137988, 808.0514327, 792.4829091, 775.7774352, 771.4592582, 777.6330473, 780.5498958, 774.9445818, 773.0120145, 783.3153164, 793.9214739, 787.9748267, 768.2767903, 751.9883636, 749.2241842, 755.4736873, 763.5043297, 768.2767903, 764.4513745, 754.2622642, 747.8600921, 749.7157042, 748.1629479, 739.1095467, 739.4124024, 753.845217, 757.6333964, 734.2986085, 705.69984, 695.9265358, 693.5024485, 688.6530327, 697.3663418, 726.8749188, 748.2386618, 738.1625018, 712.3291539, 694.7523491, 689.5243636, 694.3737794, 708.9952582, 725.8136824, 740.1323055, 755.5494012, 764.412897, 746.3073358, 705.9642182, 683.312097, 703.464417, 746.3830497, 775.3603879, 777.1787636, 760.1331976, 743.3147733, 738.6167855, 742.8220121, 740.8522085, 728.654817, 716.4201891, 717.063137, 738.0495515, 777.8986667, 820.1346327, 841.0825697, 836.4987733, 821.7258667, 811.8011345, 811.0055176
-};
-
-float electricSamples[2674] {
-  414.9802646, 415.0944567, 414.3745501, 413.3517861, 412.1391158, 411.5334012, 410.5106372, 408.2007298, 406.6467244, 405.0939602, 403.3897674, 402.1398605, 400.7757615, 399.6388055, 398.5030906, 397.8961348, 397.7447062, 398.3888985, 400.2829978, 401.8742398, 403.3140531, 404.56396, 405.4352953, 406.4195815, 406.9880595, 407.1779659, 406.6852022, 406.7224387, 406.6467244, 405.359581, 404.0327185, 402.2912891, 400.812998, 399.9416627, 398.9188988, 397.7074697, 396.6462279, 396.5705136, 396.4575627, 396.6847057, 396.9118487, 396.2676563, 396.5332771, 396.191942, 395.0562272, 394.2221284, 394.3363205, 395.8133704, 396.3061341, 397.2904203, 399.2602338, 399.8287119, 400.8514758, 401.7228111, 400.9271901, 399.5258546, 397.783184, 395.850607, 393.7293647, 391.4566939, 388.615545, 385.8885882, 383.0474394, 379.373433, 376.2666634, 373.4639923, 370.3584639, 366.948837, 363.4262592, 359.940918, 356.2669116, 352.3272845, 348.1977509, 344.4852667, 341.2283097, 338.0843036, 335.1674404, 332.6291488, 331.0379068, 330.0536206, 329.7879999, 329.0308567, 327.7809498, 328.4251421, 328.3866643, 327.3254226, 327.0610431, 326.2269444, 325.0527518, 323.3113224, 321.0374103, 318.7275029, 317.0605466, 316.7949259, 316.4163543, 316.681975, 317.5533103, 318.7647395, 321.1131246, 323.4999876, 325.6212298, 327.2881861, 329.333714, 331.3035275, 332.2878137, 333.8790557, 335.091726, 336.3416329, 338.575826, 340.3569744, 341.6825956, 342.2510737, 342.1753593, 341.1898319, 339.6755455, 337.9328749, 335.2803912, 332.7805774, 330.4321922, 328.8409503, 328.9166646, 329.902192, 331.7205769, 334.1446764, 338.0085892, 342.8195517, 346.6822233, 350.0546136, 353.9172852, 357.8196758, 361.2293027, 363.6149244, 365.1304521, 366.948837, 369.1073157, 371.1156071, 372.1756076, 371.797036, 371.0398928, 369.7515081, 367.6302659, 364.7518804, 361.6451108, 359.2594891, 357.5540551, 357.4411042, 357.5540551, 358.6525333, 360.9252042, 361.4564456, 363.4262592, 366.7974083, 368.5400789, 369.7515081, 370.2815083, 371.7597994, 373.729613, 375.0167564, 375.7751409, 375.0167564, 374.9038056, 375.2066629, 374.2211355, 372.933992, 371.34275, 369.9029367, 367.8574088, 364.789117, 362.3650175, 360.7737755, 360.0538689, 360.6980612, 360.9252042, 360.6980612, 362.1378745, 364.5247375, 367.5545515, 370.774272, 373.956756, 377.5935258, 380.2820048, 383.0102028, 386.4570663, 389.6767867, 392.2150783, 393.1621279, 393.3135566, 393.5406996, 394.4492714, 394.9420351, 394.9047986, 395.0562272, 394.0334632, 394.5249857, 395.623464, 395.5105131, 396.2304198, 396.4947993, 397.0260408, 397.4046124, 398.0488047, 400.0186183, 400.1315691, 401.4571904, 403.5784326, 403.768339, 405.0939602, 406.7609165, 408.616538, 411.0034009, 413.1246431, 415.6629347, 417.8598913, 420.4354194, 423.5794256, 426.1549537, 428.2389594, 430.2087729, 432.8600154, 435.966785, 439.3751707, 442.7847976, 446.800139, 452.671102, 459.224735, 466.2698905, 474.1119082, 481.5356354, 488.922126, 496.839858, 504.6049202, 512.0298885, 518.7349502, 524.6816275, 530.3254475, 534.9092669, 538.5832733, 541.4629, 543.3569992, 545.0227143, 546.1969069, 546.8410992, 547.0310056, 546.9552913, 547.5237693, 548.6979619, 550.5163468, 553.3190179, 557.1072164, 561.3112231, 566.2363776, 571.6902912, 576.4640171, 581.1223097, 584.9105082, 588.0929921, 590.8584266, 590.13852, 589.3056624, 589.8728992, 588.5857558, 586.8046074, 584.267557, 581.0093588, 577.941067, 574.0771541, 569.8731475, 564.3795149, 558.6984584, 553.5089244, 547.2953852, 541.6900429, 536.6519375, 532.295261, 529.4925899, 527.447062, 526.1971551, 524.3030559, 522.2947645, 520.8561924, 518.506566, 516.4238016, 513.7725591, 510.060075, 506.9160688, 503.9234913, 501.2337711, 497.408336, 492.5588958, 488.8091751, 484.073927, 479.3771566, 475.5889581, 469.5280888, 462.5586476, 455.8535859, 449.4141449, 444.0347045, 438.8836482, 434.829829, 430.6245811, 425.7006678, 421.4966611, 416.761413, 411.7233076, 406.3066306, 399.5630911, 392.5936499, 385.4343023, 378.4263833, 371.001415, 362.8193034, 355.7356701, 349.8262294, 343.6139314, 336.8716332, 329.5993347, 322.4772236, 316.2276891, 310.469677, 305.2416652, 301.0748951, 298.3479383, 297.2109823, 296.56679, 296.0367897, 296.0367897, 296.6809821, 297.7422238, 298.3479383, 298.3479383, 297.703746, 296.6052677, 295.6197403, 293.9912618, 291.9084974, 290.0516347, 287.0590572, 282.8922871, 278.5368518, 273.877318, 269.2934985, 263.4982499, 256.5275675, 251.2250825, 245.9213564, 240.656108, 237.626294, 234.5952387, 232.6254251, 233.079711, 234.670953, 237.4748653, 239.5961075, 241.678872, 244.6714495, 246.9441203, 249.028126, 250.8837475, 252.3235608, 253.649182, 255.2776606, 257.2102376, 258.1572872, 258.5358588, 257.6260457, 256.2631879, 256.9446168, 256.7174739, 255.0889954, 254.5205173, 254.4820396, 255.240424, 256.1874736, 256.6417596, 258.0815729, 260.2785294, 262.8925353, 266.3766353, 269.9376908, 273.5359829, 277.8541817, 282.5894298, 286.9461063, 291.150113, 295.0897401, 298.840702, 303.0831865, 307.0228136, 310.4311993, 314.3708264, 318.0833106, 321.4929375, 325.2041804, 328.9166646, 332.4777201, 335.7731549, 339.2200184, 343.4252662, 348.2362287, 352.9714768, 357.1754835, 361.2293027, 365.6232157, 370.205794, 374.3353276, 377.6307624, 379.9791475, 381.8360102, 383.2373458, 384.0329668, 384.562967, 384.4487749, 383.0859171, 380.6233399, 378.1607626, 376.3808554, 375.5479979, 375.3953281, 375.1309485, 375.9265695, 377.4408559, 378.1607626, 380.2447683, 382.971725, 384.8658243, 386.9870665, 388.615545, 391.0781223, 393.8435568, 396.5705136, 400.2457612, 401.2685252, 401.5713825, 403.0099546, 402.8970037, 402.5184321, 402.1398605, 402.177097, 403.2383388, 404.1469106, 405.3211032, 405.4725318, 405.0939602, 405.1696746, 404.56396, 403.8812899, 403.3512896, 402.5941464, 401.608619, 400.6243328, 400.6615694, 401.7985254, 403.0869101, 404.2226249, 405.5867239, 406.8738674, 408.5793014, 410.2834943, 411.4204503, 413.0489288, 414.9430281, 417.3671276, 418.9583695, 419.4511332, 421.7238041, 423.8065685, 425.7391455, 428.4661023, 429.7160092, 430.8144875, 431.4202021, 432.0643944, 433.2013504, 433.238587, 435.2468783, 437.8608842, 438.655264, 440.7392697, 443.1261326, 444.9060398, 446.7244247, 447.7844252, 448.3529032, 448.0115682, 447.5572822, 448.2399523, 450.1712881, 452.7095797, 454.3008217, 454.7923442, 455.5879651, 457.1407293, 458.8461634, 459.3004493, 458.3918775, 457.0277785, 455.7393938, 454.7166298, 453.8837723, 453.1638657, 452.7468163, 452.8225306, 453.5052007, 454.5279647, 455.6264429, 456.269394, 456.8006355, 457.633493, 458.1634933, 458.7704491, 459.0733064, 458.8833999, 459.3761636, 459.8689273, 460.3232133, 460.020356, 458.5048283, 457.7092074, 454.8680585, 449.8299531, 445.9660403, 440.814984, 434.6784003, 429.2232455, 423.4279969, 418.5797979, 413.9575007, 409.9049227, 408.1237743, 406.4952958, 405.132438, 405.2081523, 404.7153886, 403.1626244, 401.4571904, 399.6388055, 398.0103269, 396.7219423, 394.9805129, 394.1848919, 394.2978428, 393.1621279, 391.8365067, 390.8137428, 388.6540228, 386.0784947, 382.9344885, 379.4863838, 376.4950475, 372.2897997, 367.9716009, 364.3348311, 359.4109178, 356.1911973, 355.2441477, 353.4257627, 351.5316635, 348.8034655, 346.0765087, 343.8795522, 340.0913537, 337.6300176, 334.9402974, 330.2050493, 327.6295211, 325.8483728, 323.6514162, 321.6816027, 319.2202666, 317.3261674, 315.8478763, 314.9008267, 313.8023484, 312.0224412, 310.8482486, 310.469677, 310.923963, 311.5681553, 311.9082491, 312.7038701, 314.0679691, 315.1292108, 315.0522553, 314.3708264, 313.5379689, 313.4622545, 313.2351116, 310.7725343, 307.8556711, 305.8473798, 304.5974729, 303.120423, 300.1278455, 297.4381253, 294.8625971, 291.6428766, 288.2717275, 284.5592434, 281.6051436, 279.1040886, 276.1115111, 273.2716034, 271.5661694, 272.0204553, 272.513219, 271.2260755, 270.5434054, 270.6191197, 270.5818832, 270.8847405, 269.8992131, 268.1577837, 266.869399, 265.5437778, 264.709679, 262.8925353, 260.9984361, 259.86148, 256.8689025, 253.2706104, 249.6723183, 245.8841199, 242.2088722, 237.4363876, 232.7768538, 228.1185612, 225.0117916, 224.064742, 222.2848348, 221.0721644, 220.9207358, 220.8450215, 221.3750217, 222.0192141, 223.1561701, 224.63322, 226.7544622, 229.1785617, 231.4512325, 235.0122881, 238.6862945, 242.24735, 245.8841199, 249.1410769, 252.7021324, 255.9975672, 258.9144304, 261.4899585, 263.2711069, 266.300921, 270.4304545, 274.1801753, 279.1040886, 285.3933421, 291.5671623, 297.3251744, 301.6433731, 305.7344289, 310.0153911, 312.7795844, 314.9008267, 316.2276891, 317.4018817, 320.3559814, 324.2198942, 327.742472, 330.5836209, 333.9932478, 337.8186828, 341.7210734, 345.2051734, 345.6594593, 345.9635578, 347.0235583, 346.6822233, 346.6822233, 347.0620361, 348.3491796, 350.2060423, 352.364521, 355.8113845, 358.9938684, 363.0849242, 367.3274086, 369.5628429, 372.4039918, 376.0022838, 379.2977186, 382.5174391, 385.2828737, 388.5783085, 393.0106993, 396.1162277, 398.5403272, 402.2155748, 404.1469106, 404.9425316, 405.9268178, 405.6239605, 405.7381526, 405.9652955, 407.0637738, 409.7907306, 411.2305439, 413.2760718, 417.3671276, 421.3439913, 425.435047, 429.4503885, 433.011444, 436.8381203, 440.246506, 442.5191768, 443.9589901, 444.0731822, 443.9974679, 444.4517538, 443.6933694, 442.5576546, 441.4964129, 439.678028, 436.9138346, 433.3900156, 430.4731524, 428.4661023, 426.8376238, 424.3750465, 421.0038974, 418.6555123, 416.988556, 415.397314, 413.806072, 412.1391158, 411.8362585, 411.8747362, 411.6475933, 412.8590224, 414.2231214, 415.5487426, 417.9728421, 418.6555123, 417.2914133, 415.8900777, 414.260358, 413.1246431, 410.20778, 406.0782464, 403.199861, 398.8431845, 394.1464141, 390.3582156, 386.154209, 382.7830599, 379.9034332, 377.1379987, 374.4110419, 372.0626567, 370.205794, 368.7287441, 368.3501725, 367.3658864, 365.8503587, 363.9177817, 360.736539, 358.1225331, 356.115483, 354.524241, 351.8345208, 346.8721297, 341.7210734, 337.2886826, 333.3490554, 328.9551424, 324.5227515, 320.3559814, 316.3791177, 312.8180622, 308.8784351, 306.7944294, 305.7716655, 304.0674726, 304.2189013, 304.711665, 305.3558573, 307.5528138, 309.7497704, 311.5681553, 312.9694909, 314.2193978, 314.749398, 314.6736837, 314.3335899, 313.3865402, 313.2351116, 312.7038701, 311.265298, 309.901199, 309.0298637, 308.4241491, 307.0228136, 305.2416652, 302.779088, 300.6578457, 299.7492739, 297.703746, 296.5295534, 296.7951741, 296.7194598, 296.7194598, 296.1509818, 296.8708885, 298.1580319, 297.6280317, 297.3624109, 296.8324107, 295.9238388, 296.112504, 295.9995532, 295.4310751, 294.7111685, 293.5369759, 291.9084974, 289.1430628, 285.9233424, 283.11943, 279.862473, 276.6427525, 274.1801753, 271.6803614, 269.1805476, 266.6037783, 264.5210138, 264.1796788, 264.7481568, 265.8081573, 266.9835911, 267.3236849, 268.233498, 270.7333118, 274.1801753, 278.0813246, 282.0966661, 286.5675347, 291.264305, 295.7339324, 299.7492739, 302.7418514, 306.0000496, 309.5983417, 312.0981555, 314.0294913, 316.113497, 318.8789316, 322.2885584, 325.4325646, 329.0308567, 333.6903905, 339.4099248, 345.6594593, 351.6073778, 357.4411042, 363.2363528, 369.1830301, 375.0924708, 380.3961969, 385.3958245, 388.7682149, 390.5853586, 393.4649852, 396.7976566, 400.3214756, 403.6541469, 404.4125313, 403.6541469, 402.7070973, 402.177097, 401.1543331, 396.5332771, 391.8365067, 388.3511655, 384.1843954, 381.7975325, 379.4491473, 377.3279051, 377.213713, 377.2894273, 377.4408559, 376.3051411, 375.5852345, 377.2521908, 378.4263833, 378.3121912, 377.2894273, 376.3051411, 375.8123774, 375.5852345, 375.6609488, 375.8123774, 377.4036194, 378.2749547, 377.7064767, 378.0478117, 377.7449544, 376.418092, 374.5624705, 372.1756076, 370.2442718, 369.6385572, 370.0928431, 371.001415, 373.5397066, 376.7221905, 379.7904823, 384.1459176, 387.4798302, 389.6010724, 392.0251719, 394.9420351, 398.351662, 401.0786188, 402.2528114, 402.2155748, 402.3285257, 401.8742398, 400.9271901, 400.8514758, 400.2457612, 399.5630911, 398.4646129, 397.4046124, 397.8961348, 398.6917558, 400.3587121, 401.4571904, 400.8514758, 401.3442395, 403.0484323, 404.4882457, 404.3740536, 403.1626244, 402.2155748, 402.4427178, 402.7455751, 401.0413822, 399.4116625, 398.3888985, 396.8361343, 395.3590845, 393.1993645, 390.8137428, 388.8054514, 386.7599235, 383.8430603, 380.8504828, 378.8049549, 376.190949, 375.1681851, 375.0167564, 373.2740858, 371.8727503, 370.0928431, 368.5016012, 367.4416007, 365.7361666, 365.2818807, 363.9562595, 361.9479681, 362.4034953, 362.6306382, 361.2665392, 358.803962, 355.5842415, 353.1229054, 351.3045205, 348.6905146, 345.1679368, 340.6598317, 335.4702976, 331.5306705, 328.5008564, 325.2426582, 322.9699873, 321.4544597, 320.961696, 321.7200804, 323.121416, 326.3411365, 330.5836209, 333.8790557, 337.4401112, 341.9482164, 346.2664151, 350.2432788, 353.8043343, 356.6454832, 359.8267259, 362.4034953, 363.6149244, 364.9790234, 365.1304521, 364.1449247, 363.6149244, 361.6451108, 358.2739617, 355.0170047, 352.6301417, 350.9631855, 348.311943, 345.1294591, 341.9097386, 339.7512598, 339.371447, 338.3114465, 337.5158255, 338.0085892, 339.2957327, 340.6983095, 340.6598317, 340.6213539, 341.1525954, 341.7967877, 343.3495519, 345.0922225, 347.2891791, 349.4476578, 351.1146141, 351.796043, 351.796043, 351.9487129, 351.1530919, 351.3417571, 351.6830921, 350.5088995, 351.5316635, 353.72862, 355.7741479, 358.4638681, 361.1151106, 364.8275948, 369.5243651, 375.3196137, 382.3660105, 389.9424075, 398.351662, 407.1010104, 415.8515999, 424.2993322, 432.6328724, 441.1935556, 450.0955738, 458.7704491, 467.028275, 475.1719087, 482.1425912, 489.0747958, 495.6271876, 500.4009135, 506.3475908, 513.0141747, 520.0605715, 526.2343917, 531.6895465, 539.7959437, 546.7653849, 553.2805402, 560.4026512, 564.5309436, 569.4560981, 574.9869672, 578.8124023, 581.3506938, 582.33498, 583.5848869, 583.0548867, 580.5550729, 577.4110667, 573.5086761, 570.4403843, 567.1077129, 563.6596083, 560.4026512, 557.7886453, 555.5159745, 551.6520617, 548.0910061, 542.7885212, 536.4247946, 533.090882, 529.1140183, 524.9844848, 521.6133357, 517.8636149, 514.5669389, 510.7415039, 507.9388328, 504.8320632, 501.1580567, 498.8096716, 495.0971874, 491.5361319, 489.1877467, 486.9150758, 485.1724052, 482.6341136, 480.0970633, 476.9902937, 473.9604796, 472.7478092, 470.4751384, 468.4296105, 467.9753246, 467.2169401, 466.7254177, 467.2554179, 469.3381824, 473.2778095, 477.7859147, 482.0668768, 487.5207904, 494.0359457, 499.3781496, 503.5449197, 505.7033985, 505.6276841, 504.6049202, 502.1808207, 499.2639575, 495.5514733, 491.6118462, 488.5435544, 484.4909763, 479.6415361, 474.4520021, 468.5425614, 462.2173125, 455.5122508, 449.187002, 443.0504183, 437.7082144, 431.9514435, 424.5264752, 417.9728421, 410.776258, 402.8970037, 396.4947993, 390.5853586, 385.0172529, 378.5020977, 371.5326565, 364.0319738, 355.8498622, 349.4861356, 341.75831, 333.159149, 326.1897078, 318.5375965, 311.9839635, 305.9243353, 299.9764169, 294.1811682, 287.2104858, 282.9680014, 280.0511382, 275.3543678, 272.3245538, 271.5289328, 272.0974108, 273.081697, 272.9675049, 272.286076, 272.4759824, 273.3845543, 273.2331256, 272.5889333, 272.7031254, 273.7631259, 275.8843681, 277.8169451, 278.9154234, 280.3167589, 282.4380011, 284.5977211, 286.870392, 289.4074423, 292.2113547, 296.56679, 301.0376586, 304.3703299, 308.6512921, 314.1821612, 320.1288385, 325.5455155, 330.6593352, 337.3259191, 343.9552665, 348.7277512, 352.0244272, 354.7886205, 358.576819, 362.6306382, 365.0919743, 366.380359, 367.7816945, 370.2442718, 372.3270362, 373.2368493, 374.1454211, 374.1838989, 373.4639923, 372.6683713, 371.569893, 371.5326565, 372.4412283, 374.2596132, 376.3808554, 377.8963831, 380.6990542, 383.198868, 384.0329668, 385.0557307, 384.9043021, 384.3730606, 384.7143957, 384.9415386, 384.9415386, 382.7830599, 380.737532, 380.2820048, 378.4636199, 376.3051411, 374.3725641, 371.8727503, 369.8657002, 368.0473152, 365.2446441, 361.9107316, 358.6525333, 355.8113845, 355.5085272, 355.5085272, 353.5387136, 353.1229054, 353.7658566, 354.751384, 356.7969118, 359.1837748, 363.0092098, 366.0775017, 368.9558871, 373.8810416, 378.350669, 383.1231537, 387.3656381, 389.3739294, 391.2295509, 393.0106993, 394.2221284, 394.4120349, 394.6764144, 396.1162277, 397.8961348, 398.9188988, 398.9188988, 399.8659484, 401.9114763, 403.2383388, 403.6169104, 403.0484323, 403.3140531, 403.0484323, 401.1543331, 399.3359482, 395.3590845, 390.5096443, 386.2299233, 381.0403892, 376.190949, 371.9869424, 368.7672219, 365.1676886, 360.3579674, 355.9628131, 350.8117568, 345.8878435, 342.099645, 337.2129682, 332.4392424, 329.0308567, 326.4168508, 323.9927513, 320.5458878, 316.7949259, 314.4465407, 312.9694909, 311.6438696, 311.4167267, 311.0381551, 310.1668198, 309.5226274, 308.1970062, 306.5672864, 304.0674726, 301.4919445, 299.8249882, 296.8324107, 294.1426905, 293.0442122, 292.855547, 293.651168, 294.1811682, 294.8625971, 296.112504, 298.3851749, 300.0906089, 299.6350818, 300.8105156, 302.89328, 305.2416652, 308.4998635, 309.9384356, 311.5296775, 314.5607328, 319.6360748, 325.0527518, 327.742472, 331.8720056, 336.0387757, 338.8414468, 342.9325025, 345.2436512, 347.2134647, 351.1530919, 356.1911973, 361.8722538, 365.926073, 370.2442718, 375.3196137, 378.7292406, 381.4189609, 383.994489, 386.8741156, 390.4724077, 394.7148921, 399.1088052, 403.8812899, 408.8064444, 412.0634014, 415.170171, 417.1015068, 416.7229352, 417.1772212, 417.4428419, 417.2529355, 417.9728421, 419.7539905, 423.6936177, 427.4048606, 430.4359159, 433.8827793, 435.6254499, 437.1782141, 439.4893628, 440.4736489, 440.3606981, 439.8679344, 439.2994563, 439.0723134, 438.0123129, 435.7011643, 433.3527791, 431.1930591, 429.2232455, 428.0875307, 426.8376238, 425.2463818, 424.6406673, 425.1321898, 426.9133381, 428.8831517, 429.4888663, 430.1330586, 430.8144875, 430.3974381, 430.6245811, 431.1558226, 430.7387732, 430.2087729, 429.6787727, 430.2844873, 431.2687734, 430.8902018, 431.5716307, 433.011444, 433.6928729, 435.1326862, 436.3825932, 436.7996425, 437.0652633, 436.7239282, 436.4210709, 435.5869722, 434.4127796, 433.6556364, 431.8757292, 429.186009, 426.7234317, 424.6779038, 422.8595189, 420.2839908, 416.6099844, 412.7460715, 409.1465382, 405.132438, 400.1315691, 393.8820346, 386.722687, 380.1305762, 373.9939925, 367.2516943, 360.168061, 352.8188069, 346.0007944, 339.59859, 332.4020058, 326.0755157, 321.0758881, 315.3563538, 309.4469131, 303.5747089, 296.7951741, 289.6730631, 281.8322866, 273.9902688, 266.642256, 257.7787156, 248.9524117, 240.6188715, 231.5269469, 223.5719783, 217.0195864, 211.0716679, 205.5407988, 200.011171, 195.6172579, 191.9047737, 188.7980041, 186.5253333, 183.2298985, 179.9717002, 177.8889358, 176.6762654, 176.2976938, 176.2976938, 176.9034084, 178.2675074, 179.4801777, 181.0329419, 182.8513269, 184.2154259, 186.5253333, 189.8592458, 193.2676315, 197.3971651, 201.9425068, 206.8291835, 211.4502396, 214.4428171, 218.3452077, 223.0804558, 226.6787479, 230.5426607, 234.8608594, 239.9362014, 246.1112628, 252.2478465, 258.0815729, 264.0282501, 270.8847405, 277.5898021, 283.0821935, 287.9303925, 292.5526897, 298.499367, 304.6731872, 309.2570067, 312.4767272, 315.6592111, 320.6588387, 325.3183725, 328.2352357, 331.5691483, 335.2803912, 339.9784028, 345.0537448, 348.7662289, 352.2888067, 356.6827198, 361.53216, 365.5090237, 368.0088375, 370.0543654, 372.6311347, 374.2211355, 373.0854206, 370.9641784, 369.2215078, 368.0845518, 366.6832162, 364.4105454, 363.3505449, 362.9707321, 362.7063526, 362.7435891, 361.4179679, 360.1308244, 358.803962, 356.0397686, 352.1001415, 348.1220366, 345.7736514, 343.5009806, 341.304024, 339.6370677, 337.7814463, 338.0843036, 339.371447, 339.7884964, 340.6213539, 342.2125959, 345.0165082, 347.8191793, 350.3947074, 353.955763, 357.5540551, 361.2665392, 365.01626, 368.3501725, 371.6456073, 374.9410421, 378.0850483, 380.8504828, 383.6916317, 386.0027803, 387.9341161, 390.4724077, 392.5179356, 394.2606062, 396.0789911, 397.5932776, 399.0715687, 400.3214756, 400.8514758, 400.5486185, 399.6015689, 398.3888985, 397.4803267, 396.0789911, 393.2378423, 389.9039297, 386.1157312, 382.2518184, 378.1992404, 373.5397066, 368.8801728, 362.9707321, 357.4783407, 353.160142, 347.516322, 342.3640245, 337.9701115, 332.8947695, 327.6295211, 321.7200804, 316.0750192, 309.901199, 302.551945, 294.9383114, 286.4533426, 278.3469454, 270.9604548, 262.7783432, 253.876325, 245.6184991, 238.8762009, 233.2696175, 227.663034, 221.2235931, 215.7684383, 212.208624, 208.6090907, 205.8821339, 204.5180349, 203.7608917, 204.2151776, 205.0107986, 206.5635628, 208.9504257, 210.7688107, 213.1556736, 216.0725368, 219.1408286, 223.9505499, 230.391232, 236.9436239, 243.8001142, 251.3007969, 260.0129087, 269.5206415, 277.9683738, 286.4148648, 295.3925974, 303.4989946, 311.492441, 318.7647395, 325.0527518, 331.4164784, 337.4016334, 343.1981233, 348.2734652, 352.591664, 357.0997691, 361.1151106, 364.4490232, 368.2359804, 373.1983715, 378.2749547, 383.0474394, 387.2142095, 390.6238364, 394.6764144, 397.214706, 398.6545193, 402.0641462, 404.7153886, 406.4580592, 408.7679666, 410.2834943, 411.6091155, 412.8590224, 414.828836, 417.7456992, 420.3597051, 422.746568, 424.9807611, 427.6704813, 430.6630589, 433.0499218, 435.2084006, 437.3681205, 440.5121267, 444.4132761, 448.6942383, 453.808058, 457.7849217, 460.588834, 463.3157908, 465.8540824, 468.5810391, 471.5351389, 474.7561006, 478.3159149, 482.2183055, 486.3850756, 489.9448899, 493.8472805, 498.8853859, 503.1278703, 506.5747338, 509.9458829, 512.9769382, 516.5752302, 519.7192364, 522.3332423, 524.9472482, 525.8942978, 527.105727, 529.1512549, 529.6812551, 529.6055408, 529.8326838, 530.3254475, 531.0838319, 531.3109749, 531.2352605, 530.2882109, 529.2269692, 528.0527766, 525.9315344, 524.5301988, 523.2045776, 520.4391431, 516.7651367, 512.5983666, 508.9231189, 504.643398, 499.1125289, 493.8088027, 489.1120324, 485.3623116, 482.2183055, 478.8843929, 475.5889581, 472.5963806, 469.4523745, 465.8540824, 462.1031204, 458.7319713, 455.9665368, 452.8610084, 448.2399523, 442.6333689, 436.8753569, 431.3444877, 425.7763821, 419.2227491, 412.1018792, 405.7753891, 400.812998, 396.3061341, 391.4951717, 386.94983, 383.8058238, 382.8202964, 382.0631532, 379.9034332, 377.782191, 376.8351414, 378.1235261, 379.941911, 379.6390537, 380.3577191, 383.0859171, 385.0172529, 387.1012586, 389.7909788, 393.3135566, 398.124519, 402.0256684, 405.7381526, 409.4878733, 412.1775935, 415.5487426, 418.9956061, 421.609612, 425.0192389, 428.617531, 431.3829655, 432.8600154, 434.4127796, 437.7466922, 441.6478415, 445.5117543, 448.4670953, 450.2097659, 453.1638657, 456.6492069, 458.7319713, 459.5660701, 459.5275923, 459.9446417, 460.0575925, 458.3533997, 456.269394, 453.8837723, 450.5883375, 447.4443314, 444.640419, 441.8377479, 439.1865055, 436.193928, 433.011444, 430.0188665, 427.1777176, 424.3365688, 421.9124693, 419.4511332, 416.3828414, 413.5032148, 411.0791153, 408.7679666, 406.4580592, 403.8440533, 401.7228111, 399.8659484, 397.214706, 394.2606062, 391.1923144, 388.1997369, 385.5100166, 382.5174391, 379.1835265, 375.3580915, 371.5326565, 367.7816945, 363.8805451, 361.5693965, 359.7894894, 357.0240548, 354.7129062, 352.932999, 351.3802348, 349.8647072, 347.6677507, 345.8493657, 344.4852667, 342.8567882, 341.4169749, 340.4326887, 339.4471613, 338.1972544, 336.6444902, 334.2961051, 331.8720056, 329.4479061, 326.6067572, 323.7271306, 320.166075, 316.681975, 314.3335899, 311.6438696, 308.1970062, 305.2416652, 303.347566, 302.4762307, 301.4162302, 299.8249882, 298.8779386, 298.3479383, 298.0066033, 297.7422238, 297.0980314, 296.9466028, 297.2109823, 296.908125, 295.7339324, 294.1811682, 293.4227838, 292.855547, 292.4384976, 293.1584043, 293.651168, 293.651168, 293.3098329, 292.3243055, 291.8327831, 291.7570687, 291.6813544, 291.3400194, 290.6585905, 291.3015416, 293.3470695, 295.6197403, 297.6652682, 300.1278455, 303.4617581, 307.3256709, 310.8854852, 313.0452052, 314.9008267, 316.909118, 317.8934042, 319.0675968, 319.9016955, 319.7875034, 320.734553, 322.2885584, 323.8413226, 325.1669439, 326.1139935, 327.7052355, 329.0308567, 329.5993347, 330.6593352, 332.4392424, 335.1674404, 338.9556389, 343.0466946, 347.0992726, 351.6830921, 356.9098627, 362.4407318, 368.0473152, 373.4267557, 378.6920041, 384.79011, 390.9266936, 395.850607, 400.3214756, 404.4497679, 408.5793014, 412.7460715, 415.9273143, 419.6397984, 423.2380905, 424.8678102, 426.8748604, 429.6787727, 432.2543008, 434.7541146, 437.0267855, 439.9821265, 443.6933694, 447.59576, 451.1555743, 454.376536, 458.3161631, 462.4059777, 466.0812253, 469.3009458, 472.1420947, 474.8690515, 477.3688653, 479.4143932, 480.7784922, 481.9911625, 482.9382121, 482.8997344, 482.8240201, 482.2940198, 481.2327781, 480.4371571, 479.4143932, 478.430107, 476.9145794, 475.3233374, 474.4520021, 472.786287, 470.0965668, 466.9525606, 463.6943624, 460.5131197, 457.4448278, 454.4894869, 451.7625301, 449.8684309, 447.5200457, 444.9432763, 443.2390835, 440.5121267, 437.8981208, 435.9283072, 433.1256361, 430.1702952, 426.9505747, 423.9207606, 420.6625624, 416.7986496, 413.8445498, 411.2305439, 408.9951096, 406.9495817, 404.5254822, 402.5184321, 401.0413822, 399.6015689, 397.5932776, 395.3590845, 393.3135566, 391.2680287, 389.6010724, 387.4041159, 385.5100166, 384.6386813, 383.3502967, 381.5319117, 379.2220043, 377.5935258, 376.7594271, 375.2066629, 373.8810416, 372.9712286, 371.8727503, 371.4184644, 370.9641784, 369.9414145, 368.7672219, 367.9716009, 367.7816945, 368.0473152, 368.4258869, 369.0316014, 370.2815083, 372.1756076, 374.3725641, 376.5707619, 378.9948614, 381.9874389, 385.0557307, 388.1240225, 390.4339299, 392.2907926, 395.0177494, 396.9503264, 397.783184, 399.0715687, 400.2457612, 401.5329047, 403.199861, 404.9797681, 407.5180597, 409.980637, 412.0261649, 414.5644565, 416.9500782, 418.9583695, 421.4197056, 423.5794256, 425.1321898, 426.1934315, 426.8748604, 427.4805749, 427.3676241, 426.7619095, 426.3448601, 425.4735248, 424.6406673, 423.6551399, 421.950947, 420.3212273, 419.1470347, 418.1627486, 416.4585557, 414.2231214, 411.8362585, 409.2234938, 406.5710101, 402.5941464, 397.5175632, 393.0864136, 388.6540228, 383.6916317, 377.555048, 371.3812278, 367.2132165, 362.5164461, 357.3269121, 352.7430926, 346.7579376, 340.6213539, 334.2576273, 326.6439937, 318.50036, 310.4311993, 303.1961373, 295.5067895, 287.1732493, 279.5968523, 272.4002681, 266.4151131, 260.9599583, 254.8990889, 249.7480327, 245.6569769, 242.3603009, 238.4591515, 233.9510463, 231.3382817, 229.4441824, 227.8144627, 227.0573195, 226.1102698, 225.6932205, 226.5645558, 228.5728471, 231.2625673, 233.8380955, 237.5878162, 241.7930641, 245.3913562, 249.7480327, 254.2933744, 258.5358588, 262.8155798, 267.0593054, 271.1118834, 274.9373185, 279.1798029, 283.4235285, 287.2489636, 291.0743986, 295.3925974, 300.7335601, 305.9243353, 310.2425341, 314.1436834, 317.8561676, 321.3030311, 323.1971303, 324.3340863, 325.2426582, 325.1669439, 325.3940868, 325.6597076, 324.8640866, 323.8785592, 323.1586525, 322.6671301, 321.530174, 319.1433111, 316.7192116, 314.749398, 311.7568205, 308.0455775, 304.0674726, 299.294988, 294.1426905, 288.9159199, 283.9535288, 278.9526599, 274.3700817, 270.8090261, 266.3393988, 261.225579, 257.3989028, 253.6107043, 248.9524117, 244.1786858, 239.9362014, 236.9063873, 233.9510463, 230.391232, 227.7387484, 225.5417918, 223.4590274, 223.3448353, 223.7618847, 223.4962639, 224.1032197, 226.1487476, 228.9514187, 231.2240896, 234.102475, 238.0805799, 242.3230643, 247.8154557, 253.876325, 259.3687163, 265.4295857, 272.286076, 279.4454236, 286.4918204, 292.8927836, 299.067845, 304.9388079, 309.332721, 313.3865402, 318.9546459, 324.3340863, 329.106571, 334.3718194, 340.3569744, 346.720701, 352.591664, 359.0311049, 366.1147382, 372.8582777, 380.0933396, 387.4413524, 393.8820346, 399.8659484, 406.0410099, 412.5189286, 418.9583695, 424.7163816, 430.1330586, 435.9283072, 441.3077477, 446.0417546, 450.7012884, 454.9822506, 459.3004493, 463.8842688, 468.1267532, 471.4594246, 473.9977162, 476.3088648, 477.861629, 478.65725, 479.225728, 479.0742993, 477.9758211, 476.1946727, 473.9604796, 471.5736167, 468.7709456, 465.1714123, 461.0431199, 456.5350148, 451.3839585, 446.0802324, 440.5121267, 434.602686, 429.1475312, 423.9964749, 419.0713204, 414.3373135, 410.321972, 406.9880595, 402.7828116, 398.2759477, 394.6764144, 391.1923144, 387.1769729, 382.9344885, 378.9563836, 375.7366631, 372.8955142, 370.0928431, 367.8574088, 366.8346449, 366.6832162, 366.6459797, 366.7974083, 367.2516943, 368.1230296, 369.1073157, 370.0543654, 371.4941787, 373.6154209, 376.3808554, 379.373433, 381.8732468, 384.7143957, 388.8439292, 394.1091776, 399.8287119, 405.4352953, 410.7377802, 415.8515999, 421.5723754, 427.253432, 432.1785865, 437.5952635, 442.6333689, 446.6487104, 450.5883375, 454.4522503, 458.1262567, 461.6115979, 465.0584614, 468.1267532, 471.6108532, 475.5132438, 478.202964, 480.7027778, 483.2410694, 484.8310702, 486.1951692, 487.3321252, 487.6349825, 486.5365042, 485.2865973, 484.6796415, 483.5439267, 482.2183055, 480.2100141, 477.066008, 474.0734305, 471.422188, 468.6195169, 465.2856043, 461.5731202, 457.7092074, 452.8982449, 447.6714743, 443.3532756, 439.0723134, 434.034208, 428.9203883, 423.8450463, 418.5040836, 412.3290222, 405.7381526, 399.0715687, 391.8737433, 385.7371596, 379.714768, 372.592657, 366.1904525, 360.736539, 355.7356701, 351.3417571, 347.8948936, 345.1679368, 342.7438374, 341.3412606, 340.3184966, 339.6755455, 339.6370677, 339.2584961, 339.3342105, 340.508403, 342.326788, 343.8423156, 345.3193655, 347.8948936, 350.0546136, 352.4402353, 355.8113845, 358.6525333, 361.7965395, 365.926073, 370.5471291, 375.2438994, 379.6762902, 383.8815381, 388.2754512, 392.8965072, 397.556041, 402.2912891, 405.9652955, 408.0865377, 409.753494, 411.3819725, 412.4804508, 412.4804508, 412.2533078, 412.0261649, 411.4961646, 411.1548296, 409.7907306, 407.1779659, 404.4125313, 401.1928109, 397.328898, 393.1236502, 389.1083087, 385.3958245, 381.5703895, 378.6535263, 376.7221905, 374.4867562, 371.4184644, 367.744458, 363.5776879, 358.9181541, 353.7658566, 348.1977509, 342.2883102, 335.7359184, 329.6365713, 324.2571308, 318.8404538, 314.3335899, 310.69682, 307.5900504, 305.1287143, 303.3103294, 301.756324, 300.0906089, 298.499367, 296.9838393, 295.2039322, 293.1956408, 290.6201127, 288.1202989, 286.074771, 284.4078147, 282.9680014, 281.6051436, 280.5439019, 279.3324728, 278.1570389, 277.7784673, 277.8926594, 277.5140878, 276.9828464, 277.3626592, 278.9154234, 280.7710449, 281.9080009, 283.5749572, 286.6047712, 289.331728, 292.1728769, 295.1282179, 296.9838393, 298.7265099, 300.6963235, 302.1733734, 302.551945, 301.8705161, 301.1506094, 300.6578457, 299.5978452, 298.8022243, 299.067845, 299.2565102, 299.4464166, 300.4307028, 302.2875655, 304.1816647, 304.8630936, 304.9388079, 304.6731872, 303.4232803, 301.8332795, 299.7877517, 297.2109823, 294.7868828, 292.5899263, 290.1658268, 287.43887, 284.3321004, 281.2638086, 278.5740883, 275.0887471, 270.7705484, 266.7552069, 262.7783432, 258.9529082, 255.8461386, 253.4977534, 251.8307971, 250.6566045, 249.9366979, 250.0124122, 251.3007969, 252.2478465, 252.7021324, 253.6107043, 254.4063252, 254.9748033, 255.5805178, 256.7559516, 258.3844302, 259.2172877, 260.8085297, 263.6112008, 265.9608271, 268.006355, 271.0361691, 275.0515106, 279.6725666, 284.2563861, 288.5745848, 293.5369759, 298.5750813, 302.6276593, 306.7944294, 311.1511059, 314.7878758, 318.3861679, 322.3642728, 326.4553286, 330.6220987, 334.5989623, 338.6527816, 343.5009806, 348.0090857, 351.7203287, 355.7356701, 359.4481543, 362.4034953, 364.8275948, 366.6459797, 368.0473152, 369.9414145, 372.592657, 375.7366631, 380.5861033, 385.9643026, 390.6995507, 396.9118487, 403.7298612, 409.7907306, 415.7001713, 422.17809, 429.0718169, 434.7541146, 440.5493633, 447.2929027, 453.3910086, 459.5275923, 465.3985552, 470.4751384, 475.474766, 480.4756349, 484.9067845, 488.7706973, 492.9002309, 496.7256659, 500.1725294, 504.3020629, 507.8246407, 510.7415039, 514.0754164, 517.1052305, 519.4163791, 520.9691433, 522.5603853, 525.0229625, 526.6129633, 526.7271554, 527.6357272, 528.2042052, 527.1442048, 526.1214408, 525.3258198, 523.9617208, 521.3861927, 517.7879006, 514.453988, 511.0071246, 506.6504481, 501.3479632, 495.475759, 489.3391753, 482.8612566, 476.4602934, 470.4751384, 464.9442693, 459.4146414, 453.4667229, 447.59576, 442.0648909, 436.8381203, 432.1028722, 428.2389594, 425.3978105, 423.2765683, 421.7238041, 420.6625624, 420.1697987, 420.0568478, 420.0940844, 420.7382767, 422.8595189, 426.8376238, 431.9129658, 437.7082144, 444.5647047, 452.3682447, 460.8532135, 469.1110394, 477.9758211, 488.5063178, 499.0740511, 508.9988333, 518.506566, 526.8028697, 533.9622173, 540.2502296, 545.0611921, 548.0910061, 549.340913, 549.1137701, 548.2051982, 546.1211926, 541.9929002, 536.6891741, 530.7424968, 523.8102922, 516.0067522, 507.9388328, 500.0211007, 491.8774669, 483.0139265, 474.1491448, 465.9297967, 458.087779, 449.9056674, 441.156319, 432.8984932, 425.5877169, 418.5797979, 410.8134945, 403.1626244, 397.5175632, 391.7980289, 385.0172529, 379.0320979, 373.0854206, 366.8346449, 361.2293027, 355.5470049, 349.4104213, 342.8567882, 336.1517265, 330.8107638, 325.6597076, 318.9546459, 313.1593973, 308.764243, 303.916044, 299.2565102, 295.8096467, 293.8398332, 292.5899263, 292.0214483, 292.7041184, 294.1054539, 297.1737457, 301.4919445, 306.4158578, 313.4237768, 321.6816027, 329.9779063, 337.7814463, 346.1907008, 356.0770052, 363.993496, 370.7370355, 379.0320979, 386.3813519, 392.139364, 398.2759477, 404.6769109, 410.7377802, 415.7386491, 419.7155127, 424.7536182, 429.4131519, 431.8757292, 434.7156369, 437.3296428, 437.8608842, 438.0880272, 438.1637415, 437.4053571, 435.398307, 432.4057295, 429.2989599, 424.9435246, 418.9956061, 412.5561651, 405.2838666, 397.1389916, 388.8054514, 380.2062905, 370.8884641, 360.5466326, 350.4331852, 341.1141176, 330.2422858, 319.0303602, 308.5371, 295.6582181, 281.6808579, 269.1048333, 256.9830946, 244.6342129, 231.4897103, 218.3824442, 206.1477546, 193.3818236, 179.9717002, 168.3054887, 158.6078495, 148.5701164, 138.9854281, 130.9932229, 124.326639, 118.4171983, 112.7349006, 109.8937517, 110.2723233, 110.8792791, 112.5462354, 117.1672914, 124.2124469, 132.4702728, 141.1066703, 150.767073, 162.9260482, 178.3816995, 194.0260159, 208.3062334, 225.201698, 243.7628776, 260.884244, 278.8397091
-};
-
-#endif // Samples_H
