+++ PGDrive/platformio.ini
@@ -1,0 +1,20 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:teensy31]
+platform = teensy
+board = teensy31
+framework = arduino
+lib_extra_dirs = ~/Documents/Arduino/libraries
+
+[env:teensy35]
+platform = Teensy
+board = teensy35
+framework = arduino
